Engine-neutral omnibar shared by WK and CEF panes; extension popovers
OmnibarHostingPanel protocol + OmnibarPaneChrome shared chrome view:
the omnibar reducer, suggestion building, native field, suggestions
list, and chrome style move verbatim into Sources/Panels/Omnibar/;
BrowserPanelView re-adopts them through accessory slots with the
WKWebView-specific blur handoff moved into BrowserPanel (behavior
unchanged by construction). CEFBrowserPanel conforms: full omnibar
with history-backed suggestions (same profile-scoped store, visits
recorded on load-end, typed navigations boost frecency), Cmd+L
routing, and the suggestions dropdown mounted as an AppKit sibling
above the CEF view in CEFBrowserHostView (portal file untouched).

Extension popup action bar in the CEF pane chrome: one button per
staged extension (Chromium unpacked-id derivation, manifest name/icon
with __MSG_ localization), click opens the popup page in a transient
NSPopover hosting a CEF browser with async teardown on dismiss — no
Chrome-window popups. en+ja strings; CEFOmnibarIntegrationTests wired.

Implemented-by: codex
