+++ .github/swift-file-length-budget.tsv
@@ -11,7 +11,7 @@
 11344	Sources/Panels/BrowserPanel.swift
 9497	cmuxTests/CLINotifyProcessIntegrationRegressionTests.swift
 8016	CLI/cmux_open.swift
-7959	Sources/Panels/BrowserPanelView.swift
+4283	Sources/Panels/BrowserPanelView.swift
 7709	Packages/iOS/CmuxMobileShell/Sources/CmuxMobileShell/MobileShellComposite.swift
 7489	cmuxTests/WorkspaceUnitTests.swift
 7312	cmuxTests/WorkspaceRemoteConnectionTests.swift
