+++ Sources/Panels/BrowserPanelView.swift
@@ -9,7 +9,6 @@ import AppKit
 import ObjectiveC
 
 private var cmuxBrowserPanelNeedsRenderingStateReattachKey: UInt8 = 0
-let browserOmnibarTextFieldIdentifier = NSUserInterfaceItemIdentifier("cmux.browserOmnibarTextField")
 
 private func browserPanelViewObjectID(_ object: AnyObject?) -> String {
     guard let object else { return "nil" }
@@ -284,91 +283,7 @@ enum BrowserProfilePopoverDebugSettings {
     }
 }
 
-struct OmnibarInlineCompletion: Equatable {
-    let typedText: String
-    let displayText: String
-    let acceptedText: String
 
-    var suffixRange: NSRange {
-        let typedCount = typedText.utf16.count
-        let fullCount = displayText.utf16.count
-        return NSRange(location: typedCount, length: max(0, fullCount - typedCount))
-    }
-}
-
-func resolvedBrowserChromeBackgroundColor(
-    for colorScheme: ColorScheme,
-    themeBackgroundColor: NSColor,
-    drawsBackground: Bool
-) -> NSColor {
-    guard drawsBackground else { return .clear }
-    switch colorScheme {
-    case .dark, .light:
-        return themeBackgroundColor
-    @unknown default:
-        return themeBackgroundColor
-    }
-}
-
-func resolvedBrowserChromeColorScheme(
-    for colorScheme: ColorScheme,
-    themeBackgroundColor: NSColor,
-    windowBackgroundColor: NSColor = .windowBackgroundColor
-) -> ColorScheme {
-    let perceivedBackgroundColor = themeBackgroundColor.alphaComponent < 0.999
-        ? cmuxCompositedNSColor(themeBackgroundColor, over: windowBackgroundColor)
-        : themeBackgroundColor
-    return cmuxReadableColorScheme(for: perceivedBackgroundColor)
-}
-
-func resolvedBrowserOmnibarPillBackgroundColor(
-    for colorScheme: ColorScheme,
-    themeBackgroundColor: NSColor
-) -> NSColor {
-    let darkenMix: CGFloat
-    switch colorScheme {
-    case .light:
-        darkenMix = 0.04
-    case .dark:
-        darkenMix = 0.05
-    @unknown default:
-        darkenMix = 0.04
-    }
-
-    let blendedColor = themeBackgroundColor.blended(withFraction: darkenMix, of: .black) ?? themeBackgroundColor
-    return blendedColor.withAlphaComponent(themeBackgroundColor.alphaComponent)
-}
-
-private struct BrowserChromeStyle {
-    let backgroundColor: NSColor
-    let colorScheme: ColorScheme
-    let omnibarPillBackgroundColor: NSColor
-
-    static func resolve(
-        for colorScheme: ColorScheme,
-        themeBackgroundColor: NSColor,
-        drawsBackground: Bool
-    ) -> BrowserChromeStyle {
-        let backgroundColor = resolvedBrowserChromeBackgroundColor(
-            for: colorScheme,
-            themeBackgroundColor: themeBackgroundColor,
-            drawsBackground: drawsBackground
-        )
-        let chromeColorScheme = resolvedBrowserChromeColorScheme(
-            for: colorScheme,
-            themeBackgroundColor: themeBackgroundColor
-        )
-        let omnibarPillBackgroundColor = resolvedBrowserOmnibarPillBackgroundColor(
-            for: chromeColorScheme,
-            themeBackgroundColor: themeBackgroundColor
-        )
-        return BrowserChromeStyle(
-            backgroundColor: backgroundColor,
-            colorScheme: chromeColorScheme,
-            omnibarPillBackgroundColor: omnibarPillBackgroundColor
-        )
-    }
-}
 
 /// View for rendering a browser panel with address bar
 struct BrowserPanelView: View {
@@ -391,12 +306,6 @@ struct BrowserPanelView: View {
     /// through this rather than the former `BrowserInstalledBrowserDetector` static
     /// namespace.
     private let installedBrowserDetector = BrowserInstalledBrowserDetector()
-    @State private var omnibarState = OmnibarState()
-    @State private var addressBarFocused: Bool = false
-    @AppStorage(BrowserSearchSettingsStore.searchEngineKey) private var searchEngineRaw = BrowserSearchSettingsStore.defaultSearchEngine.rawValue
-    @AppStorage(BrowserSearchSettingsStore.customSearchEngineNameKey) private var customSearchEngineName = BrowserSearchSettingsStore.defaultCustomSearchEngineName
-    @AppStorage(BrowserSearchSettingsStore.customSearchEngineURLTemplateKey) private var customSearchEngineURLTemplate = BrowserSearchSettingsStore.defaultCustomSearchEngineURLTemplate
-    @AppStorage(BrowserSearchSettingsStore.searchSuggestionsEnabledKey) private var searchSuggestionsEnabledStorage = BrowserSearchSettingsStore.defaultSearchSuggestionsEnabled
     @AppStorage(BrowserDevToolsButtonDebugSettings.iconNameKey) private var devToolsIconNameRaw = BrowserDevToolsButtonDebugSettings.defaultIcon.rawValue
     @AppStorage(BrowserDevToolsButtonDebugSettings.iconColorKey) private var devToolsIconColorRaw = BrowserDevToolsButtonDebugSettings.defaultColor.rawValue
     @AppStorage(BrowserToolbarAccessorySpacingDebugSettings.key) private var browserToolbarAccessorySpacingRaw = BrowserToolbarAccessorySpacingDebugSettings.defaultSpacing
@@ -410,40 +319,20 @@ struct BrowserPanelView: View {
     @AppStorage(BrowserImportHintSettings.dismissedKey) private var isBrowserImportHintDismissed = BrowserImportHintSettings.defaultDismissed
     @ObservedObject private var keyboardShortcutSettingsObserver = KeyboardShortcutSettingsObserver.shared
     @LiveSetting(\.shortcuts.showModifierHoldHints) private var showModifierHoldHints
-    @State private var omnibarSuggestionRefreshScheduler = OmnibarSuggestionRefreshScheduler()
-    @State private var omnibarSuggestionRefreshConsumerTask: Task<Void, Never>?
-    @State private var suggestionTask: Task<Void, Never>?
-    @State private var isLoadingRemoteSuggestions: Bool = false
-    @State private var latestRemoteSuggestionQuery: String = ""
-    @State private var latestRemoteSuggestions: [String] = []
     @State private var emptyStateImportBrowsers: [InstalledBrowserCandidate] = []
     @State private var emptyStateImportBrowserRefreshTask: Task<Void, Never>?
     @State private var emptyStateImportBrowserRefreshGeneration: UInt64 = 0
-    @State private var inlineCompletion: OmnibarInlineCompletion?
     @State private var screenshotPageCopied: Bool = false
     @State private var screenshotPageCaptureInProgress: Bool = false
     @State private var screenshotPageCopiedTimer: Timer?
-    @State private var omnibarSelectionRange: NSRange = NSRange(location: NSNotFound, length: 0)
-    @State private var omnibarHasMarkedText: Bool = false
-    @State private var suppressNextFocusLostRevert: Bool = false
     @State private var focusFlashOpacity: Double = 0.0
     @State private var focusFlashAnimationGeneration: Int = 0
-    @State private var omnibarPillFrame: CGRect = .zero
-    @State private var addressBarHeight: CGFloat = 0
-    @State private var addressBarWidth: CGFloat = 0
+    @State private var sharedOmnibarSuggestions: BrowserPortalOmnibarSuggestionsConfiguration?
+    @State private var sharedAddressBarFocused = false
+    @State private var sharedAddressBarHeight: CGFloat = 0
 
-    /// Below this chrome width the full accessory row would crowd out the
-    /// omnibar, so the plain-action buttons collapse into an overflow menu.
-    private static let compactChromeWidthThreshold: CGFloat = 420
-
-    private var isChromeCompact: Bool {
-        addressBarWidth > 0 && addressBarWidth < Self.compactChromeWidthThreshold
-    }
     @State private var isBrowserImportHintPopoverPresented = false
     @State private var focusModeShortcutHintMonitor = WindowScopedShortcutHintModifierMonitor(activation: .commandOnly)
-    @State private var lastHandledAddressBarFocusRequestId: UUID?
-    @State private var omnibarSelectAllRequestId: UInt64 = 0
-    @State private var pendingFocusGainedSelectionIntent: BrowserAddressBarFocusSelectionIntent = .preserveFieldEditorSelection
     @State private var isBrowserProfileMenuPresented = false
     @State private var isBrowserThemeMenuPresented = false
     @State private var browserChromeStyle: BrowserChromeStyle
@@ -457,10 +346,6 @@ struct BrowserPanelView: View {
     // appearance view-state seed (the empty-state import list) so a spurious appear
     // does no work. App-once settings work lives in the model bootstrap, not here.
     @State private var didCompleteInitialBrowserPanelSetup = false
-    // Keep this below half of the compact omnibar height so it reads as a squircle,
-    // not a capsule.
-    private let omnibarPillCornerRadius: CGFloat = 10
-    private let addressBarVerticalPadding: CGFloat = 4
     // Toolbar/omnibar sizes derived from the tab bar font size. Names and call
     // sites are unchanged; the values now scale via `chromeMetrics`.
     private var chromeMetrics: BrowserChromeMetrics {
@@ -493,37 +378,6 @@ struct BrowserPanelView: View {
         ))
     }
 
-    private var searchConfiguration: BrowserSearchConfiguration {
-        BrowserSearchSettingsStore().configuration(
-            engineRaw: searchEngineRaw,
-            customName: customSearchEngineName,
-            customURLTemplate: customSearchEngineURLTemplate
-        )
-    }
-
-    private var searchSuggestionsEnabled: Bool {
-        // Touch @AppStorage so SwiftUI invalidates this view when settings change.
-        _ = searchSuggestionsEnabledStorage
-        return BrowserSearchSettingsStore(defaults: .standard).currentSearchSuggestionsEnabled
-    }
-
-    private var remoteSuggestionsEnabled: Bool {
-        // Deterministic UI-test hook: force remote path on even if a persisted
-        // setting disabled suggestions in previous sessions.
-        if ProcessInfo.processInfo.environment["CMUX_UI_TEST_REMOTE_SUGGESTIONS_JSON"] != nil ||
-            UserDefaults.standard.string(forKey: "CMUX_UI_TEST_REMOTE_SUGGESTIONS_JSON") != nil {
-            return true
-        }
-        // Keep UI tests deterministic by disabling network suggestions when requested.
-        if ProcessInfo.processInfo.environment["CMUX_UI_TEST_DISABLE_REMOTE_SUGGESTIONS"] == "1" {
-            return false
-        }
-        return searchSuggestionsEnabled
-    }
-
-    private var hasActionableOmnibarSuggestions: Bool {
-        !omnibarHasMarkedText && !omnibarState.suggestions.isEmpty
-    }
 
     private var devToolsIconOption: BrowserDevToolsIconOption {
         BrowserDevToolsIconOption(rawValue: devToolsIconNameRaw) ?? BrowserDevToolsButtonDebugSettings.defaultIcon
@@ -569,63 +423,21 @@ struct BrowserPanelView: View {
         browserChromeStyle.backgroundColor
     }
 
-    private var browserChromeColorScheme: ColorScheme {
-        browserChromeStyle.colorScheme
-    }
-
     private var browserContentAccessibilityIdentifier: String {
         "BrowserPanelContent.\(panel.id.uuidString)"
     }
 
-    private var omnibarPillBackgroundColor: NSColor {
-        browserChromeStyle.omnibarPillBackgroundColor
-    }
-
-    private var hasVisibleOmnibarSuggestions: Bool {
-        panel.isOmnibarVisible && addressBarFocused && hasActionableOmnibarSuggestions && omnibarPillFrame.width > 0
-    }
-
     private var shouldRenderOmnibarSuggestionsInPortal: Bool {
-        hasVisibleOmnibarSuggestions &&
-            panel.shouldRenderWebView
+        sharedOmnibarSuggestions != nil && panel.shouldRenderWebView
     }
 
     private var shouldRenderOmnibarSuggestionsInSwiftUI: Bool {
-        hasVisibleOmnibarSuggestions && !shouldRenderOmnibarSuggestionsInPortal
-    }
-
-    private var omnibarSuggestionsFrameInPortal: CGRect? {
-        guard shouldRenderOmnibarSuggestionsInPortal else { return nil }
-        let top = max(0, omnibarPillFrame.maxY + 3 - addressBarHeight)
-        let height = OmnibarSuggestionsView.popupHeight(for: omnibarState.suggestions)
-        guard omnibarPillFrame.width > 0, height > 0 else { return nil }
-        return CGRect(
-            x: omnibarPillFrame.minX,
-            y: top,
-            width: omnibarPillFrame.width,
-            height: height
-        )
+        sharedOmnibarSuggestions != nil && !shouldRenderOmnibarSuggestionsInPortal
     }
 
     private var portalOmnibarSuggestions: BrowserPortalOmnibarSuggestionsConfiguration? {
-        guard let frame = omnibarSuggestionsFrameInPortal else { return nil }
-        return BrowserPortalOmnibarSuggestionsConfiguration(
-            panelId: panel.id,
-            popupFrame: frame,
-            colorScheme: browserChromeColorScheme,
-            engineName: searchConfiguration.displayName,
-            items: omnibarState.suggestions,
-            selectedIndex: omnibarState.selectedSuggestionIndex,
-            isLoadingRemoteSuggestions: isLoadingRemoteSuggestions,
-            searchSuggestionsEnabled: remoteSuggestionsEnabled,
-            onCommit: { item in
-                commitSuggestion(item)
-            },
-            onHighlight: { idx in
-                let effects = omnibarReduce(state: &omnibarState, event: .highlightIndex(idx))
-                applyOmnibarEffects(effects)
-            }
-        )
+        guard shouldRenderOmnibarSuggestionsInPortal else { return nil }
+        return sharedOmnibarSuggestions
     }
 
     private var developerToolsButtonHelp: String {
@@ -810,7 +622,6 @@ struct BrowserPanelView: View {
         // re-running this per appear is harmless once the heavy/one-time work is
         // gated out.
         performInitialBrowserPanelSetupIfNeeded()
-        startOmnibarSuggestionRefreshConsumer()
         refreshBrowserChromeStyle()
         panel.noteWebViewVisibility(
             isVisibleInUI && isCurrentPaneOwner,
@@ -818,12 +629,7 @@ struct BrowserPanelView: View {
         )
         panel.refreshAppearanceDrivenColors()
         panel.setBrowserThemeMode(browserThemeMode)
-        applyPendingAddressBarFocusRequestIfNeeded()
-        syncURLFromPanel()
-        // If the browser surface is focused but has no URL loaded yet, auto-focus the omnibar.
-        autoFocusOmnibarIfBlank()
         syncWebViewResponderPolicyWithViewState(reason: "onAppear")
-        panel.historyStore.loadIfNeeded()
 #if DEBUG
         logBrowserFocusState(event: "view.onAppear")
 #endif
@@ -848,19 +654,11 @@ struct BrowserPanelView: View {
     }
 
     private func handleOmnibarVisibilityChange(_ isVisible: Bool) {
-        if !isVisible {
-            hideSuggestions()
-            setAddressBarFocused(false, reason: "omnibarVisibility.hidden")
-            addressBarHeight = 0
-        } else {
-            applyPendingAddressBarFocusRequestIfNeeded()
-        }
+        _ = isVisible
         syncWebViewResponderPolicyWithViewState(reason: "omnibarVisibilityChanged")
     }
 
     private func handleBrowserPanelDisappear() {
-        stopOmnibarSuggestionRefreshConsumer()
-        cancelPendingOmnibarSuggestionWork()
         focusModeShortcutHintMonitor.stop()
         screenshotPageCopiedTimer?.invalidate()
         screenshotPageCopiedTimer = nil
@@ -876,44 +674,18 @@ struct BrowserPanelView: View {
         cmuxDebugLog(
             "browser.focus.clickIntent panel=\(panel.id.uuidString.prefix(5)) " +
             "isFocused=\(isFocused ? 1 : 0) " +
-            "addressFocused=\(addressBarFocused ? 1 : 0)"
+            "addressFocused=\(sharedAddressBarFocused ? 1 : 0)"
         )
 #endif
-        if addressBarFocused {
-#if DEBUG
-            logBrowserFocusState(event: "addressBarFocus.webViewClickBlur")
-#endif
-            setAddressBarFocused(false, reason: "webView.clickIntent")
-        }
+        NotificationCenter.default.post(name: .browserDidBlurAddressBar, object: panel.id)
         if !isFocused {
             onRequestPanelFocus()
         }
     }
 
-    private func handleExternalAddressBarBlur(_ notification: Notification) {
-        guard let panelId = notification.object as? UUID,
-              panelId == panel.id,
-              addressBarFocused else {
-            return
-        }
-#if DEBUG
-        logBrowserFocusState(event: "addressBarFocus.externalBlur")
-#endif
-        setAddressBarFocused(false, reason: "notification.externalBlur")
-    }
 
     private func handleCurrentURLChange() {
         refreshBrowserChromeStyle()
-        let addressWasEmpty = omnibarState.buffer.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-        syncURLFromPanel()
-        // If we auto-focused a blank omnibar but then a URL loads programmatically, move focus
-        // into WebKit unless the user had already started typing.
-        if addressBarFocused,
-           !panel.shouldSuppressWebViewFocus(),
-           addressWasEmpty,
-           !isBrowserContentBlankForOmnibar() {
-            setAddressBarFocused(false, reason: "panel.currentURL.loaded")
-        }
         if panel.isShowingNewTabPage {
             refreshEmptyStateImportBrowsers()
         }
@@ -943,16 +715,8 @@ struct BrowserPanelView: View {
         panel.refreshAppearanceDrivenColors()
     }
 
-    private func handleCommandPaletteVisibilityChange(_ notification: Notification) {
-        guard commandPaletteVisibilityNotificationMatchesPanelWindow(notification) else { return }
-        applyPendingAddressBarFocusRequestIfNeeded()
-    }
-
     private func handleProfileChange() {
         panel.historyStore.loadIfNeeded()
-        if addressBarFocused {
-            refreshSuggestions()
-        }
     }
 
     private func handlePanelVisibilityChange(_ visibleInUI: Bool) {
@@ -979,14 +743,9 @@ struct BrowserPanelView: View {
         )
 #endif
         // Ensure this view doesn't retain focus while hidden (bonsplit keepAllAlive).
-        if focused {
-            applyPendingAddressBarFocusRequestIfNeeded()
-            autoFocusOmnibarIfBlank()
-        } else {
+        if !focused {
             panel.invalidateAddressBarPageFocusRestoreAttempts()
             panel.clearBrowserFocusMode(reason: "panelFocus.onChange.unfocused")
-            hideSuggestions()
-            setAddressBarFocused(false, reason: "panelFocus.onChange.unfocused")
             // Surface switches in split layouts can keep the browser visible, so
             // `isVisibleInUI` never flips to false. Check for an attached-inspector
             // X-close when focus leaves as well so the persisted intent stays in sync.
@@ -999,70 +758,6 @@ struct BrowserPanelView: View {
             reason: "panelFocusChanged",
             isPanelFocusedOverride: focused
         )
-    }
-
-    private func handleAddressBarFocusedChange(_ focused: Bool) {
-#if DEBUG
-        logBrowserFocusState(
-            event: "addressBarFocus.onChange",
-            detail: "next=\(focused ? 1 : 0)"
-        )
-#endif
-        let urlString = panel.preferredURLStringForOmnibar() ?? ""
-        if focused {
-            let selectionIntent = pendingFocusGainedSelectionIntent
-            pendingFocusGainedSelectionIntent = .preserveFieldEditorSelection
-            panel.beginSuppressWebViewFocusForAddressBar()
-            NotificationCenter.default.post(name: .browserDidFocusAddressBar, object: panel.id)
-            // Only request panel focus if this pane isn't currently focused. When already
-            // focused (e.g. Cmd+L), forcing focus can steal first responder back to WebKit.
-            if !isFocused {
-#if DEBUG
-                logBrowserFocusState(event: "addressBarFocus.requestPanelFocus")
-#endif
-                onRequestPanelFocus()
-            }
-            let effects = omnibarReduce(
-                state: &omnibarState,
-                event: .focusGained(currentURLString: urlString, shouldSelectAll: selectionIntent.shouldSelectAll)
-            )
-            applyOmnibarEffects(effects)
-            refreshInlineCompletion()
-        } else {
-            pendingFocusGainedSelectionIntent = .preserveFieldEditorSelection
-            panel.endSuppressWebViewFocusForAddressBar()
-            NotificationCenter.default.post(name: .browserDidBlurAddressBar, object: panel.id)
-            if suppressNextFocusLostRevert {
-                suppressNextFocusLostRevert = false
-                let effects = omnibarReduce(state: &omnibarState, event: .focusLostPreserveBuffer(currentURLString: urlString))
-                applyOmnibarEffects(effects)
-            } else {
-                let effects = omnibarReduce(state: &omnibarState, event: .focusLostRevertBuffer(currentURLString: urlString))
-                applyOmnibarEffects(effects)
-            }
-            inlineCompletion = nil
-        }
-        syncWebViewResponderPolicyWithViewState(reason: "addressBarFocusChanged")
-#if DEBUG
-        logBrowserFocusState(event: "addressBarFocus.onChange.applied")
-#endif
-    }
-
-    private func handleMoveOmnibarSelection(_ notification: Notification) {
-        guard let panelId = notification.object as? UUID, panelId == panel.id else { return }
-        guard canHandleOmnibarSuggestionInteraction() else { return }
-        guard let delta = notification.userInfo?["delta"] as? Int, delta != 0 else { return }
-#if DEBUG
-        logBrowserFocusState(event: "addressBarFocus.moveSelection", detail: "delta=\(delta)")
-#endif
-        let effects = omnibarReduce(state: &omnibarState, event: .moveSelection(delta: delta))
-        applyOmnibarEffects(effects)
-        refreshInlineCompletion()
-    }
-
-    private func handleHistoryEntriesChange() {
-        guard addressBarFocused else { return }
-        refreshSuggestions()
     }
 
     @ViewBuilder
@@ -1096,34 +791,88 @@ struct BrowserPanelView: View {
 
     @ViewBuilder
     private var omnibarSuggestionsOverlayView: some View {
-        if shouldRenderOmnibarSuggestionsInSwiftUI {
+        if shouldRenderOmnibarSuggestionsInSwiftUI,
+           let configuration = sharedOmnibarSuggestions {
             OmnibarSuggestionsView(
-                engineName: searchConfiguration.displayName,
-                items: omnibarState.suggestions,
-                selectedIndex: omnibarState.selectedSuggestionIndex,
-                isLoadingRemoteSuggestions: isLoadingRemoteSuggestions,
-                searchSuggestionsEnabled: remoteSuggestionsEnabled,
-                onCommit: { item in
-                    commitSuggestion(item)
-                },
-                onHighlight: { idx in
-                    let effects = omnibarReduce(state: &omnibarState, event: .highlightIndex(idx))
-                    applyOmnibarEffects(effects)
-                }
+                engineName: configuration.engineName,
+                items: configuration.items,
+                selectedIndex: configuration.selectedIndex,
+                isLoadingRemoteSuggestions: configuration.isLoadingRemoteSuggestions,
+                searchSuggestionsEnabled: configuration.searchSuggestionsEnabled,
+                onCommit: configuration.onCommit,
+                onHighlight: configuration.onHighlight
             )
-            .frame(width: omnibarPillFrame.width)
-            .offset(x: omnibarPillFrame.minX, y: omnibarPillFrame.maxY + 3)
+            .frame(width: configuration.popupFrame.width)
+            .offset(x: configuration.popupFrame.minX, y: configuration.popupFrame.minY)
             .zIndex(1000)
-            .environment(\.colorScheme, browserChromeColorScheme)
+            .environment(\.colorScheme, configuration.colorScheme)
         }
     }
 
     @ViewBuilder
     private var omnibarHeaderView: some View {
-        if panel.isOmnibarVisible {
-            addressBar
-                .fixedSize(horizontal: false, vertical: true)
-        }
+        OmnibarPaneChrome(
+            panel: panel,
+            isFocused: isFocused,
+            chromeStyle: browserChromeStyle,
+            tabBarFontSize: tabBarFontSize,
+            accessorySpacing: browserToolbarAccessorySpacing,
+            onRequestPanelFocus: onRequestPanelFocus,
+            onReloadOrStop: handleReloadOrStopButtonAction,
+            onReload: handleReloadButtonContextMenuAction,
+            onHardReload: handleHardRefreshButtonAction,
+            onAddressBarFocusStateChange: { focused in
+                sharedAddressBarFocused = focused
+                syncWebViewResponderPolicyWithViewState(reason: "sharedOmnibar.focusChanged")
+            },
+            onChromeHeightChange: { height in
+                sharedAddressBarHeight = height
+            },
+            onSuggestionsPresentationChange: { configuration in
+                sharedOmnibarSuggestions = configuration
+            },
+            leadingAccessories: {
+                BrowserPDFDocumentToolbarButtons(
+                    panel: panel,
+                    iconPointSize: chromeMetrics.navigationIconFontSize,
+                    hitSize: addressBarButtonHitSize
+                )
+
+                if panel.isDownloading || !panel.recentDownloads.isEmpty {
+                    BrowserDownloadsToolbarButton(
+                        downloads: panel.recentDownloads,
+                        isDownloading: panel.isDownloading,
+                        iconPointSize: chromeMetrics.navigationIconFontSize,
+                        hitSize: addressBarButtonHitSize,
+                        onOpen: { panel.openDownload($0) },
+                        onReveal: { panel.revealDownloadInFinder($0) },
+                        onClear: { panel.clearRecentDownloads() }
+                    )
+                }
+            },
+            trailingAccessories: { isCompact in
+                if shouldShowToolbarImportHintChip {
+                    browserImportHintToolbarChip
+                }
+                if isCompact {
+                    browserOverflowMenu
+                    browserProfileButton
+                    browserThemeModeButton
+                } else {
+                    browserFocusModeButtonWithShortcutHint
+                    screenshotPageButton
+                    reactGrabButton
+                    browserProfileButton
+                    browserThemeModeButton
+                    developerToolsButton
+                }
+            }
+        )
+        .background(
+            WindowAccessor(refreshID: showModifierHoldHints) { window in
+                focusModeShortcutHintMonitor.setHostWindow(showModifierHoldHints ? window : nil)
+            }
+        )
     }
 
     private var browserPanelBaseView: some View {
@@ -1132,25 +881,16 @@ struct BrowserPanelView: View {
         VStack(spacing: 0) {
             omnibarHeaderView
             webView
+                .overlay(omnibarSuggestionsOverlayView, alignment: .topLeading)
         }
         .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
         .overlay(browserFindOverlayView)
         .overlay(focusFlashOverlayView)
-        .overlay(omnibarSuggestionsOverlayView, alignment: .topLeading)
     }
 
     private var browserPanelLifecycleView: some View {
         browserPanelBaseView
         .coordinateSpace(name: "BrowserPanelViewSpace")
-        .onPreferenceChange(OmnibarPillFramePreferenceKey.self) { frame in
-            omnibarPillFrame = frame
-        }
-        .onPreferenceChange(BrowserAddressBarHeightPreferenceKey.self) { height in
-            addressBarHeight = height
-        }
-        .onPreferenceChange(BrowserAddressBarWidthPreferenceKey.self) { width in
-            addressBarWidth = width
-        }
         .onReceive(NotificationCenter.default.publisher(for: .webViewDidReceiveClick)) { notification in
             handleBrowserWebViewClickIntent(notification)
         }
@@ -1181,9 +921,6 @@ struct BrowserPanelView: View {
         .onChange(of: colorScheme) { _ in
             handleSystemColorSchemeChange()
         }
-        .onChange(of: panel.pendingAddressBarFocusRequestId) { _ in
-            applyPendingAddressBarFocusRequestIfNeeded()
-        }
         .onChange(of: panel.isOmnibarVisible) { _, isVisible in
             handleOmnibarVisibilityChange(isVisible)
         }
@@ -1194,9 +931,6 @@ struct BrowserPanelView: View {
 
     var body: some View {
         browserPanelLifecycleView
-        .onReceive(NotificationCenter.default.publisher(for: .commandPaletteVisibilityDidChange)) { notification in
-            handleCommandPaletteVisibilityChange(notification)
-        }
         .onChange(of: panel.profileID) { _ in
             handleProfileChange()
         }
@@ -1206,150 +940,11 @@ struct BrowserPanelView: View {
         .onChange(of: isFocused) { focused in
             handlePanelFocusChange(focused)
         }
-        .onChange(of: addressBarFocused) { focused in
-            handleAddressBarFocusedChange(focused)
-        }
-        .onReceive(NotificationCenter.default.publisher(for: .browserMoveOmnibarSelection)) { notification in
-            handleMoveOmnibarSelection(notification)
-        }
-        .onReceive(panel.historyStore.$entries) { _ in
-            handleHistoryEntriesChange()
-        }
-        .onReceive(NotificationCenter.default.publisher(for: .browserDidBlurAddressBar)) { notification in
-            handleExternalAddressBarBlur(notification)
-        }
         .onReceive(NotificationCenter.default.publisher(for: .ghosttyDefaultBackgroundDidChange)) { _ in
             refreshBrowserChromeStyle()
         }
     }
 
-    private var addressBar: some View {
-        HStack(spacing: 8) {
-            addressBarButtonBar
-
-            omnibarField
-                .accessibilityIdentifier("BrowserOmnibarPill")
-                .accessibilityLabel("Browser omnibar")
-
-            HStack(spacing: browserToolbarAccessorySpacing) {
-                if shouldShowToolbarImportHintChip {
-                    browserImportHintToolbarChip
-                }
-                if isChromeCompact {
-                    // Narrow panes can't fit the full accessory row without
-                    // clipping the omnibar; collapse the plain-action buttons
-                    // into an overflow menu and keep the popover-anchored
-                    // profile/theme controls present.
-                    browserOverflowMenu
-                    browserProfileButton
-                    browserThemeModeButton
-                } else {
-                    browserFocusModeButtonWithShortcutHint
-                    screenshotPageButton
-                    reactGrabButton
-                    browserProfileButton
-                    browserThemeModeButton
-                    developerToolsButton
-                }
-            }
-        }
-        .padding(.horizontal, 8)
-        .padding(.vertical, addressBarVerticalPadding)
-        .background(browserChromeBackground)
-        .background {
-            GeometryReader { geo in
-                Color.clear.preference(
-                    key: BrowserAddressBarWidthPreferenceKey.self,
-                    value: geo.size.width
-                )
-            }
-        }
-        .background(
-            WindowAccessor(refreshID: showModifierHoldHints) { window in
-                focusModeShortcutHintMonitor.setHostWindow(showModifierHoldHints ? window : nil)
-            }
-        )
-        .background {
-            GeometryReader { geo in
-                Color.clear
-                    .preference(
-                        key: BrowserAddressBarHeightPreferenceKey.self,
-                        value: geo.size.height
-                    )
-            }
-        }
-        // Keep the omnibar stack above WKWebView so the suggestions popup is visible.
-        .zIndex(1)
-        .environment(\.colorScheme, browserChromeColorScheme)
-    }
-
-    private var addressBarButtonBar: some View {
-        return HStack(spacing: 0) {
-            Button(action: {
-                #if DEBUG
-                cmuxDebugLog("browser.back panel=\(panel.id.uuidString.prefix(5))")
-                #endif
-                panel.goBack()
-            }) {
-                CmuxSystemSymbolImage(systemName: "chevron.left", pointSize: chromeMetrics.navigationIconFontSize, weight: .medium)
-                    .frame(width: addressBarButtonHitSize, height: addressBarButtonHitSize, alignment: .center)
-                    .contentShape(Rectangle())
-            }
-            .buttonStyle(OmnibarAddressButtonStyle())
-            .disabled(!panel.canGoBack)
-            .opacity(panel.canGoBack ? 1.0 : 0.4)
-            .safeHelp(String(localized: "browser.goBack", defaultValue: "Go Back"))
-
-            Button(action: {
-                #if DEBUG
-                cmuxDebugLog("browser.forward panel=\(panel.id.uuidString.prefix(5))")
-                #endif
-                panel.goForward()
-            }) {
-                CmuxSystemSymbolImage(systemName: "chevron.right", pointSize: chromeMetrics.navigationIconFontSize, weight: .medium)
-                    .frame(width: addressBarButtonHitSize, height: addressBarButtonHitSize, alignment: .center)
-                    .contentShape(Rectangle())
-            }
-            .buttonStyle(OmnibarAddressButtonStyle())
-            .disabled(!panel.canGoForward)
-            .opacity(panel.canGoForward ? 1.0 : 0.4)
-            .safeHelp(String(localized: "browser.goForward", defaultValue: "Go Forward"))
-
-            Button(action: handleReloadOrStopButtonAction) {
-                CmuxSystemSymbolImage(systemName: panel.isLoading ? "xmark" : "arrow.clockwise", pointSize: chromeMetrics.navigationIconFontSize, weight: .medium)
-                    .frame(width: addressBarButtonHitSize, height: addressBarButtonHitSize, alignment: .center)
-                    .contentShape(Rectangle())
-            }
-            .buttonStyle(OmnibarAddressButtonStyle())
-            .contextMenu {
-                Button(String(localized: "browser.reload", defaultValue: "Reload")) {
-                    handleReloadButtonContextMenuAction()
-                }
-                Button(String(localized: "menu.view.hardRefresh", defaultValue: "Hard Refresh")) {
-                    handleHardRefreshButtonAction()
-                }
-            }
-            .safeHelp(panel.isLoading ? String(localized: "browser.stop", defaultValue: "Stop") : String(localized: "browser.reload", defaultValue: "Reload"))
-
-            BrowserPDFDocumentToolbarButtons(
-                panel: panel,
-                iconPointSize: chromeMetrics.navigationIconFontSize,
-                hitSize: addressBarButtonHitSize
-            )
-
-            if panel.isDownloading || !panel.recentDownloads.isEmpty {
-                BrowserDownloadsToolbarButton(
-                    downloads: panel.recentDownloads,
-                    isDownloading: panel.isDownloading,
-                    iconPointSize: chromeMetrics.navigationIconFontSize,
-                    hitSize: addressBarButtonHitSize,
-                    onOpen: { panel.openDownload($0) },
-                    onReveal: { panel.revealDownloadInFinder($0) },
-                    onClear: { panel.clearRecentDownloads() }
-                )
-            }
-        }
-    }
 
     private var screenshotPageButton: some View {
         Button(action: handleScreenshotPageButtonAction) {
@@ -1696,99 +1291,6 @@ struct BrowserPanelView: View {
         devToolsColorOption.color
     }
 
-    private var omnibarField: some View {
-        let showSecureBadge = panel.currentURL?.scheme == "https"
-
-        return HStack(spacing: 4) {
-            if showSecureBadge {
-                CmuxSystemSymbolImage(systemName: "lock.fill", pointSize: chromeMetrics.secureBadgeFontSize)
-                    .foregroundColor(.secondary)
-            }
-
-            OmnibarTextFieldRepresentable(
-                panelId: panel.id,
-                fontSize: chromeMetrics.omnibarFontSize,
-                text: Binding(
-                    get: { omnibarState.buffer },
-                    set: { newValue in
-                        let effects = omnibarReduce(state: &omnibarState, event: .bufferChanged(newValue))
-                        applyOmnibarEffects(effects)
-                        if !effects.shouldClearInlineCompletion {
-                            refreshInlineCompletion()
-                        }
-                    }
-                ),
-                isFocused: $addressBarFocused,
-                selectAllRequestId: omnibarSelectAllRequestId,
-                inlineCompletion: inlineCompletion,
-                placeholder: String(localized: "browser.addressBar.placeholder", defaultValue: "Search or enter URL"),
-                onTap: {
-                    handleOmnibarTap()
-                },
-                onSubmit: { liveField in
-                    handleOmnibarSubmit(liveField: liveField)
-                },
-                onEscape: {
-                    handleOmnibarEscape()
-                },
-                onFieldLostFocus: {
-                    setAddressBarFocused(false, reason: "omnibar.fieldLostFocus")
-                },
-                onMoveSelection: { delta in
-                    guard canHandleOmnibarSuggestionInteraction() else { return }
-                    let effects = omnibarReduce(state: &omnibarState, event: .moveSelection(delta: delta))
-                    applyOmnibarEffects(effects)
-                    refreshInlineCompletion()
-                },
-                onDeleteSelectedSuggestion: {
-                    deleteSelectedSuggestionIfPossible()
-                },
-                onAcceptInlineCompletion: {
-                    acceptInlineCompletion()
-                },
-                onDeleteBackwardWithInlineSelection: {
-                    handleInlineBackspace()
-                },
-                onClearTypedPrefixWithInlineSelection: {
-                    handleInlineClearTypedPrefix()
-                },
-                onDeleteWordBackwardWithInlineSelection: {
-                    handleInlineDeleteWordBackward()
-                },
-                onSelectionChanged: { selectionRange, hasMarkedText in
-                    handleOmnibarSelectionChange(range: selectionRange, hasMarkedText: hasMarkedText)
-                },
-                shouldSuppressWebViewFocus: {
-                    panel.shouldSuppressWebViewFocus()
-                }
-            )
-                .frame(height: chromeMetrics.omnibarFieldHeight)
-                .accessibilityIdentifier("BrowserOmnibarTextField")
-        }
-        .padding(.horizontal, 8)
-        .padding(.vertical, 4)
-        .background(
-            RoundedRectangle(cornerRadius: omnibarPillCornerRadius, style: .continuous)
-                .fill(Color(nsColor: omnibarPillBackgroundColor))
-        )
-        .overlay {
-            BrowserOmnibarInteractionRepresentable(panelId: panel.id)
-        }
-        .overlay(
-            RoundedRectangle(cornerRadius: omnibarPillCornerRadius, style: .continuous)
-                .stroke(addressBarFocused ? cmuxAccentColor() : Color.clear, lineWidth: 1)
-        )
-        .accessibilityElement(children: .contain)
-        .background {
-            GeometryReader { geo in
-                Color.clear
-                    .preference(
-                        key: OmnibarPillFramePreferenceKey.self,
-                        value: geo.frame(in: .named("BrowserPanelViewSpace"))
-                    )
-            }
-        }
-    }
 
     private var webView: some View {
         // Canvas-hosted panes force inline hosting: window-portal positioning
@@ -1803,7 +1305,7 @@ struct BrowserPanelView: View {
                     paneId: paneId,
                     shouldAttachWebView: isVisibleInUI && isCurrentPaneOwner && !useLocalInlineDeveloperToolsHosting,
                     useLocalInlineHosting: useLocalInlineDeveloperToolsHosting,
-                    shouldFocusWebView: isFocused && !addressBarFocused,
+                    shouldFocusWebView: isFocused && !sharedAddressBarFocused,
                     isPanelFocused: isFocused,
                     portalZPriority: portalPriority,
                     paneDropZone: paneDropZone,
@@ -1823,7 +1325,7 @@ struct BrowserPanelView: View {
                         )
                     },
                     omnibarSuggestions: portalOmnibarSuggestions,
-                    paneTopChromeHeight: panel.isOmnibarVisible ? addressBarHeight : 0
+                    paneTopChromeHeight: panel.isOmnibarVisible ? sharedAddressBarHeight : 0
                 )
                 .accessibilityIdentifier("BrowserWebViewSurface")
                 // Keep the host stable for normal pane churn, but force a remount when
@@ -1835,11 +1337,11 @@ struct BrowserPanelView: View {
                 .simultaneousGesture(TapGesture().onEnded {
                     // Chrome-like behavior: clicking web content while editing the
                     // omnibar should commit blur and revert transient edits.
-                    if addressBarFocused {
-#if DEBUG
-                        logBrowserFocusState(event: "webContent.tapBlur")
-#endif
-                        setAddressBarFocused(false, reason: "webContent.tapBlur")
+                    if sharedAddressBarFocused {
+                        NotificationCenter.default.post(
+                            name: .browserDidBlurAddressBar,
+                            object: panel.id
+                        )
                     }
                 })
             } else {
@@ -1848,8 +1350,11 @@ struct BrowserPanelView: View {
                     .accessibilityIdentifier(browserContentAccessibilityIdentifier)
                     .onTapGesture {
                         onRequestPanelFocus()
-                        if addressBarFocused {
-                            setAddressBarFocused(false, reason: "placeholderContent.tapBlur")
+                        if sharedAddressBarFocused {
+                            NotificationCenter.default.post(
+                                name: .browserDidBlurAddressBar,
+                                object: panel.id
+                            )
                         }
                     }
                     .overlay(alignment: .topLeading) {
@@ -1950,67 +1455,6 @@ struct BrowserPanelView: View {
         cmuxWebView.allowsFirstResponderAcquisition = next
     }
 
-    private func canHandleOmnibarSelectionNavigation() -> Bool {
-        if addressBarFocused {
-            return true
-        }
-        if AppDelegate.shared?.focusedBrowserAddressBarPanelId() == panel.id {
-            return true
-        }
-        let fieldWindow = panel.webView.window ?? NSApp.keyWindow ?? NSApp.mainWindow
-        if let field = browserOmnibarField(panelId: panel.id, in: fieldWindow),
-           field.currentEditor() != nil {
-            return true
-        }
-        return false
-    }
-
-    private func canHandleOmnibarSuggestionInteraction() -> Bool {
-        canHandleOmnibarSelectionNavigation() && hasActionableOmnibarSuggestions
-    }
-
-    private func setAddressBarFocused(
-        _ focused: Bool,
-        reason: String,
-        focusGainedSelectionIntent: BrowserAddressBarFocusSelectionIntent = .preserveFieldEditorSelection
-    ) {
-#if DEBUG
-        if addressBarFocused == focused {
-            logBrowserFocusState(
-                event: "addressBarFocus.write.noop",
-                detail: "reason=\(reason) value=\(focused ? 1 : 0)"
-            )
-        } else {
-            logBrowserFocusState(
-                event: "addressBarFocus.write",
-                detail: "reason=\(reason) old=\(addressBarFocused ? 1 : 0) new=\(focused ? 1 : 0)"
-            )
-        }
-#endif
-        if focused, !addressBarFocused {
-            pendingFocusGainedSelectionIntent = focusGainedSelectionIntent
-        } else if !focused {
-            pendingFocusGainedSelectionIntent = .preserveFieldEditorSelection
-        }
-        addressBarFocused = focused
-        if focused {
-            panel.noteAddressBarFocused()
-        }
-    }
-
-    private func browserFocusResponderChainContains(
-        _ start: NSResponder?,
-        target: NSResponder
-    ) -> Bool {
-        var current = start
-        var hops = 0
-        while let responder = current, hops < 64 {
-            if responder === target { return true }
-            current = responder.nextResponder
-            hops += 1
-        }
-        return false
-    }
 
     private func isPanelFocusedInModel() -> Bool {
         if let paneOwnershipOverride { return paneOwnershipOverride && isFocused }
@@ -2025,182 +1469,7 @@ struct BrowserPanelView: View {
         isPanelFocusedInModel() && panel.canApplySearchFocusRequest(generation)
     }
 
-    private func shouldApplyAddressBarExitFallback(in window: NSWindow) -> Bool {
-        // Navigation-triggered omnibar blur can still be unwinding when Cmd+F opens
-        // the browser find bar. Once find is visible, any delayed omnibar-exit
-        // handoff must not reclaim first responder for WebKit.
-        panel.webView.window === window &&
-            isPanelFocusedInModel() &&
-            panel.searchState == nil
-    }
 
-#if DEBUG
-    private func browserFocusWindow() -> NSWindow? {
-        panel.webView.window ?? NSApp.keyWindow ?? NSApp.mainWindow
-    }
-
-    private func browserFocusResponderDescription(_ responder: NSResponder?) -> String {
-        guard let responder else { return "nil" }
-        return String(describing: type(of: responder))
-    }
-
-    private func logBrowserFocusState(event: String, detail: String = "") {
-        let window = browserFocusWindow()
-        let firstResponder = window?.firstResponder
-        let firstResponderType = browserFocusResponderDescription(firstResponder)
-        let webResponder = browserFocusResponderChainContains(firstResponder, target: panel.webView) ? 1 : 0
-        var line =
-            "browser.focus.trace event=\(event) panel=\(panel.id.uuidString.prefix(5)) " +
-            "panelFocused=\(isFocused ? 1 : 0) addrFocused=\(addressBarFocused ? 1 : 0) " +
-            "suppressWeb=\(panel.shouldSuppressWebViewFocus() ? 1 : 0) " +
-            "suppressAuto=\(panel.shouldSuppressOmnibarAutofocus() ? 1 : 0) " +
-            "webResponder=\(webResponder) win=\(window?.windowNumber ?? -1) fr=\(firstResponderType)"
-        if let pending = panel.pendingAddressBarFocusRequestId {
-            line += " pending=\(pending.uuidString.prefix(8))"
-        }
-        if !detail.isEmpty {
-            line += " \(detail)"
-        }
-        cmuxDebugLog(line)
-    }
-#endif
-
-    private func syncURLFromPanel() {
-        let urlString = panel.preferredURLStringForOmnibar() ?? ""
-        let effects = omnibarReduce(state: &omnibarState, event: .panelURLChanged(currentURLString: urlString))
-        applyOmnibarEffects(effects)
-    }
-
-    private func isCommandPaletteVisibleForPanelWindow() -> Bool {
-        guard let app = AppDelegate.shared else { return false }
-
-        if let window = panel.webView.window, app.isCommandPaletteVisible(for: window) {
-            return true
-        }
-
-        if let manager = app.tabManagerFor(tabId: panel.workspaceId),
-           let windowId = app.windowId(for: manager),
-           let window = app.mainWindow(for: windowId),
-           app.isCommandPaletteVisible(for: window) {
-            return true
-        }
-
-        if let keyWindow = NSApp.keyWindow, app.isCommandPaletteVisible(for: keyWindow) {
-            return true
-        }
-        if let mainWindow = NSApp.mainWindow, app.isCommandPaletteVisible(for: mainWindow) {
-            return true
-        }
-        return false
-    }
-
-    private func commandPaletteVisibilityNotificationMatchesPanelWindow(_ notification: Notification) -> Bool {
-        if let notificationWindow = notification.object as? NSWindow,
-           panel.webView.window === notificationWindow {
-            return true
-        }
-
-        guard let app = AppDelegate.shared,
-              let manager = app.tabManagerFor(tabId: panel.workspaceId),
-              let panelWindowId = app.windowId(for: manager) else {
-            return false
-        }
-
-        if let notificationWindowId = notification.userInfo?["windowId"] as? UUID {
-            return notificationWindowId == panelWindowId
-        }
-
-        if let notificationWindow = notification.object as? NSWindow,
-           let panelWindow = app.mainWindow(for: panelWindowId) {
-            return notificationWindow === panelWindow
-        }
-        return false
-    }
-
-    private func applyPendingAddressBarFocusRequestIfNeeded() {
-        guard let requestId = panel.pendingAddressBarFocusRequestId else {
-            return
-        }
-        guard panel.panelType == .browser else {
-            lastHandledAddressBarFocusRequestId = requestId
-            panel.acknowledgeAddressBarFocusRequest(requestId)
-#if DEBUG
-            logBrowserFocusState(
-                event: "addressBarFocus.request.apply.skip",
-                detail: "reason=chrome_hidden request=\(requestId.uuidString.prefix(8))"
-            )
-#endif
-            return
-        }
-        guard !isCommandPaletteVisibleForPanelWindow() else {
-#if DEBUG
-            logBrowserFocusState(
-                event: "addressBarFocus.request.apply.skip",
-                detail: "reason=command_palette_visible request=\(requestId.uuidString.prefix(8))"
-            )
-#endif
-            return
-        }
-        guard lastHandledAddressBarFocusRequestId != requestId else {
-#if DEBUG
-            logBrowserFocusState(
-                event: "addressBarFocus.request.apply.skip",
-                detail: "reason=already_handled request=\(requestId.uuidString.prefix(8))"
-            )
-#endif
-            return
-        }
-        lastHandledAddressBarFocusRequestId = requestId
-        let selectionIntent = panel.pendingAddressBarFocusSelectionIntent
-        panel.beginSuppressWebViewFocusForAddressBar()
-#if DEBUG
-        logBrowserFocusState(
-            event: "addressBarFocus.request.apply",
-            detail: "request=\(requestId.uuidString.prefix(8)) selection=\(String(describing: selectionIntent))"
-        )
-#endif
-
-        if addressBarFocused {
-            // Re-run explicit selection behavior only for requests that own it
-            // (Cmd+L), without replacing a caret from focus restoration.
-            let effects = omnibarReduce(
-                state: &omnibarState,
-                event: .focusReasserted(
-                    shouldSelectAll: browserOmnibarShouldSelectAllOnFocusReassertion(
-                        selectionIntent: selectionIntent
-                    )
-                )
-            )
-            applyOmnibarEffects(effects)
-            refreshInlineCompletion()
-#if DEBUG
-            logBrowserFocusState(
-                event: "addressBarFocus.request.apply",
-                detail: "request=\(requestId.uuidString.prefix(8)) mode=refresh"
-            )
-#endif
-        } else {
-            setAddressBarFocused(
-                true,
-                reason: "request.apply",
-                focusGainedSelectionIntent: selectionIntent
-            )
-#if DEBUG
-            logBrowserFocusState(
-                event: "addressBarFocus.request.apply",
-                detail: "request=\(requestId.uuidString.prefix(8)) mode=set_focused"
-            )
-#endif
-        }
-
-        panel.acknowledgeAddressBarFocusRequest(requestId)
-#if DEBUG
-        logBrowserFocusState(
-            event: "addressBarFocus.request.ack",
-            detail: "request=\(requestId.uuidString.prefix(8))"
-        )
-#endif
-    }
 
     private var emptyBrowserStateCardOverlay: some View {
         VStack {
@@ -2353,60 +1622,6 @@ struct BrowserPanelView: View {
     }
 
     /// Treat content as blank only if neither WebKit nor the panel model has a nonblank URL.
-    private func isBrowserContentBlankForOmnibar() -> Bool {
-        panel.preferredURLStringForOmnibar() == nil
-    }
-
-    private func autoFocusOmnibarIfBlank() {
-        guard panel.isOmnibarVisible else {
-#if DEBUG
-            logBrowserFocusState(event: "addressBarFocus.autoFocus.skip", detail: "reason=omnibar_hidden")
-#endif
-            return
-        }
-        guard isFocused else {
-#if DEBUG
-            logBrowserFocusState(event: "addressBarFocus.autoFocus.skip", detail: "reason=panel_not_focused")
-#endif
-            return
-        }
-        guard !addressBarFocused else {
-#if DEBUG
-            logBrowserFocusState(event: "addressBarFocus.autoFocus.skip", detail: "reason=already_focused")
-#endif
-            return
-        }
-        guard !isCommandPaletteVisibleForPanelWindow() else {
-#if DEBUG
-            logBrowserFocusState(event: "addressBarFocus.autoFocus.skip", detail: "reason=command_palette_visible")
-#endif
-            return
-        }
-        // If a test/automation explicitly focused WebKit, don't steal focus back.
-        guard !panel.shouldSuppressOmnibarAutofocus() else {
-#if DEBUG
-            logBrowserFocusState(event: "addressBarFocus.autoFocus.skip", detail: "reason=autofocus_suppressed")
-#endif
-            return
-        }
-        // If a real navigation is underway (e.g. open_browser https://...), don't steal focus.
-        guard !panel.webView.isLoading else {
-#if DEBUG
-            logBrowserFocusState(event: "addressBarFocus.autoFocus.skip", detail: "reason=webview_loading")
-#endif
-            return
-        }
-        guard isBrowserContentBlankForOmnibar() else {
-#if DEBUG
-            logBrowserFocusState(event: "addressBarFocus.autoFocus.skip", detail: "reason=webview_not_blank")
-#endif
-            return
-        }
-        setAddressBarFocused(true, reason: "autoFocus.blank")
-#if DEBUG
-        logBrowserFocusState(event: "addressBarFocus.autoFocus.apply")
-#endif
-    }
 
     private func refreshEmptyStateImportBrowsers() {
         emptyStateImportBrowserRefreshTask?.cancel()
@@ -2450,173 +1665,6 @@ struct BrowserPanelView: View {
         panel.setBrowserThemeMode(mode)
     }
 
-    private func handleOmnibarTap() {
-#if DEBUG
-        logBrowserFocusState(event: "addressBar.tap")
-#endif
-        let wasAddressBarFocused = addressBarFocused
-        let shouldRequestPanelFocus = !isFocused
-        if !wasAddressBarFocused {
-            // Mark focused before pane selection converges so WebKit focus is not
-            // briefly re-acquired during `focusPane`.
-            setAddressBarFocused(true, reason: "omnibar.tap")
-        }
-        if shouldRequestPanelFocus && wasAddressBarFocused {
-            onRequestPanelFocus()
-        }
-    }
-
-    private func hideSuggestions() {
-        cancelPendingOmnibarSuggestionWork()
-        let effects = omnibarReduce(state: &omnibarState, event: .suggestionsUpdated([]))
-        applyOmnibarEffects(effects)
-        inlineCompletion = nil
-    }
-
-    private func startOmnibarSuggestionRefreshConsumer() {
-        guard omnibarSuggestionRefreshConsumerTask == nil else { return }
-        let scheduler = omnibarSuggestionRefreshScheduler
-        omnibarSuggestionRefreshConsumerTask = Task { @MainActor in
-            for await generation in scheduler.refreshStream {
-                guard scheduler.shouldProcessRefresh(generation) else { continue }
-                refreshSuggestions()
-            }
-        }
-    }
-
-    private func stopOmnibarSuggestionRefreshConsumer() {
-        omnibarSuggestionRefreshConsumerTask?.cancel()
-        omnibarSuggestionRefreshConsumerTask = nil
-    }
-
-    private func cancelPendingOmnibarSuggestionWork() {
-        omnibarSuggestionRefreshScheduler.cancelPendingRefresh()
-        suggestionTask?.cancel()
-        suggestionTask = nil
-        isLoadingRemoteSuggestions = false
-    }
-
-    private func handleOmnibarSubmit(liveField: OmnibarLiveFieldSnapshot?) {
-        let decision = omnibarSubmitDecision(
-            liveField: liveField,
-            state: omnibarState,
-            inlineCompletion: inlineCompletion,
-            canInteractWithSuggestions: canHandleOmnibarSuggestionInteraction()
-        )
-        switch decision {
-        case .commitSelectedSuggestion:
-            commitSelectedSuggestion()
-        case .navigate(let text):
-            if text != omnibarState.buffer {
-                // Reconcile the reducer with the live field before navigating so
-                // blur and URL-change handling see the text that was submitted.
-                let effects = omnibarReduce(state: &omnibarState, event: .bufferChanged(text))
-                applyOmnibarEffects(effects)
-            }
-            panel.navigateSmart(text)
-            hideSuggestions()
-            suppressNextFocusLostRevert = true
-            setAddressBarFocused(false, reason: "omnibar.submit.navigate")
-        }
-    }
-
-    private func commitSelectedSuggestion() {
-        let idx = omnibarState.selectedSuggestionIndex
-        guard idx >= 0, idx < omnibarState.suggestions.count else { return }
-        commitSuggestion(omnibarState.suggestions[idx])
-    }
-
-    private func commitSuggestion(_ suggestion: OmnibarSuggestion) {
-        // Treat this as a commit, not a user edit: don't refetch suggestions while we're navigating away.
-        omnibarState.buffer = suggestion.completion
-        omnibarState.isUserEditing = false
-        switch suggestion.kind {
-        case .switchToTab(let tabId, let panelId, _, _):
-            AppDelegate.shared?.tabManager?.focusTab(tabId, surfaceId: panelId)
-        default:
-            panel.navigateSmart(suggestion.completion)
-        }
-        hideSuggestions()
-        inlineCompletion = nil
-        suppressNextFocusLostRevert = true
-        setAddressBarFocused(false, reason: "suggestion.commit")
-    }
-
-    private func handleOmnibarEscape() {
-        guard addressBarFocused else { return }
-
-        // Chrome-like flow: clear inline completion first, then apply normal escape behavior.
-        if inlineCompletion != nil {
-            inlineCompletion = nil
-            return
-        }
-
-        let effects = omnibarReduce(state: &omnibarState, event: .escape)
-        applyOmnibarEffects(effects)
-        refreshInlineCompletion()
-    }
-
-    private func handleOmnibarSelectionChange(range: NSRange, hasMarkedText: Bool) {
-        let didBeginComposition = !omnibarHasMarkedText && hasMarkedText
-        omnibarSelectionRange = range
-        omnibarHasMarkedText = hasMarkedText
-        if didBeginComposition {
-            hideSuggestions()
-        } else {
-            // Do not refresh suggestions from selection-state publication. On
-            // composition end, the committed buffer change immediately follows.
-            refreshInlineCompletion()
-        }
-    }
-
-    private func acceptInlineCompletion() {
-        guard let completion = inlineCompletion else { return }
-        let effects = omnibarReduce(state: &omnibarState, event: .bufferChanged(completion.displayText))
-        applyOmnibarEffects(effects)
-        inlineCompletion = nil
-    }
-
-    private func handleInlineBackspace() {
-        guard let completion = inlineCompletion else { return }
-        let prefix = completion.typedText
-        guard !prefix.isEmpty else { return }
-        let updated = String(prefix.dropLast())
-        let effects = omnibarReduce(state: &omnibarState, event: .bufferChanged(updated))
-        applyOmnibarEffects(effects)
-        omnibarSelectionRange = NSRange(location: updated.utf16.count, length: 0)
-        if !effects.shouldClearInlineCompletion {
-            refreshInlineCompletion()
-        }
-    }
-
-    private func handleInlineClearTypedPrefix() {
-        guard inlineCompletion != nil else { return }
-        // Modified Backspace dismisses the current inline suggestion instead of
-        // refetching suggestions for the shorter prefix.
-        _ = omnibarReduce(state: &omnibarState, event: .bufferChanged(""))
-        omnibarSelectionRange = NSRange(location: 0, length: 0)
-        hideSuggestions()
-    }
-
-    private func handleInlineDeleteWordBackward() {
-        guard let completion = inlineCompletion else { return }
-        let updated = omnibarPrefixAfterDeletingTrailingWord(from: completion.typedText)
-        // Modified Backspace dismisses the current inline suggestion instead of
-        // refetching suggestions for the shorter prefix.
-        _ = omnibarReduce(state: &omnibarState, event: .bufferChanged(updated))
-        omnibarSelectionRange = NSRange(location: updated.utf16.count, length: 0)
-        hideSuggestions()
-    }
-
-    private func deleteSelectedSuggestionIfPossible() {
-        let idx = omnibarState.selectedSuggestionIndex
-        guard idx >= 0, idx < omnibarState.suggestions.count else { return }
-
-        let target = omnibarState.suggestions[idx]
-        guard case .history(let url, _) = target.kind else { return }
-        guard panel.historyStore.removeHistoryEntry(urlString: url) else { return }
-        refreshSuggestions()
-    }
 
     private func applyBrowserProfileSelection(_ profileID: UUID) {
         isBrowserProfileMenuPresented = false
@@ -2681,2731 +1729,10 @@ struct BrowserPanelView: View {
         guard alert.runModal() == .alertFirstButtonReturn else { return }
         _ = browserProfileStore.renameProfile(id: profile.id, to: input.stringValue)
     }
-
-    private func refreshInlineCompletion() {
-        inlineCompletion = omnibarInlineCompletionForDisplay(
-            typedText: omnibarState.buffer,
-            suggestions: omnibarState.suggestions,
-            isFocused: addressBarFocused,
-            selectionRange: omnibarSelectionRange,
-            hasMarkedText: omnibarHasMarkedText
-        )
-    }
-
-    private func refreshSuggestions() {
-#if DEBUG
-        let typingTimingStart = CmuxTypingTiming.start()
-        defer {
-            let trimmedQuery = omnibarState.buffer.trimmingCharacters(in: .whitespacesAndNewlines)
-            CmuxTypingTiming.logDuration(
-                path: "browser.omnibar.refreshSuggestions",
-                startedAt: typingTimingStart,
-                extra: "focused=\(addressBarFocused ? 1 : 0) queryLen=\(trimmedQuery.utf8.count) suggestionCount=\(omnibarState.suggestions.count)"
-            )
-        }
-#endif
-        suggestionTask?.cancel()
-        suggestionTask = nil
-        isLoadingRemoteSuggestions = false
-
-        guard addressBarFocused, !omnibarHasMarkedText else {
-#if DEBUG
-            cmuxDebugLog(
-                "browser.omnibar.suggestions refresh=skip " +
-                "panel=\(panel.id.uuidString.prefix(5)) " +
-                "focused=\(addressBarFocused ? 1 : 0) marked=\(omnibarHasMarkedText ? 1 : 0) " +
-                "bufferLen=\(omnibarState.buffer.utf8.count)"
-            )
-#endif
-            let effects = omnibarReduce(state: &omnibarState, event: .suggestionsUpdated([]))
-            applyOmnibarEffects(effects)
-            return
-        }
-
-        let query = omnibarState.buffer.trimmingCharacters(in: .whitespacesAndNewlines)
-        let historyEntries: [BrowserHistoryStore.Entry] = {
-            if query.isEmpty {
-                return panel.historyStore.recentSuggestions(limit: 12)
-            }
-            return panel.historyStore.suggestions(for: query, limit: 12)
-        }()
-        let openTabMatches = query.isEmpty ? [] : matchingOpenTabSuggestions(for: query, limit: 12)
-        let isSingleCharacterQuery = omnibarSingleCharacterQuery(for: query) != nil
-        let remoteSuggestionsEngine = searchConfiguration.remoteSuggestionsEngine
-        let allowsRemoteSuggestions = remoteSuggestionsEnabled && remoteSuggestionsEngine != nil
-        if !allowsRemoteSuggestions {
-            latestRemoteSuggestionQuery = ""
-            latestRemoteSuggestions = []
-        }
-        let staleRemote: [String]
-        if query.isEmpty || isSingleCharacterQuery {
-            staleRemote = []
-        } else {
-            staleRemote = staleRemoteSuggestionsForDisplay(
-                query: query,
-                allowsRemoteSuggestions: allowsRemoteSuggestions
-            )
-        }
-        let resolvedURL = query.isEmpty ? nil : panel.resolveNavigableURL(from: query)
-        let items = buildOmnibarSuggestions(
-            query: query,
-            engineName: searchConfiguration.displayName,
-            historyEntries: historyEntries,
-            openTabMatches: openTabMatches,
-            remoteQueries: staleRemote,
-            resolvedURL: resolvedURL,
-            limit: 8
-        )
-        let effects = omnibarReduce(state: &omnibarState, event: .suggestionsUpdated(items))
-        applyOmnibarEffects(effects)
-        refreshInlineCompletion()
-#if DEBUG
-        cmuxDebugLog(
-            "browser.omnibar.suggestions refresh=local " +
-            "panel=\(panel.id.uuidString.prefix(5)) queryLen=\(query.utf8.count) " +
-            "items=\(items.count) history=\(historyEntries.count) openTabs=\(openTabMatches.count) " +
-            "staleRemote=\(staleRemote.count) frameWidth=\(String(format: "%.1f", omnibarPillFrame.width)) " +
-            "portal=\(shouldRenderOmnibarSuggestionsInPortal ? 1 : 0)"
-        )
-#endif
-
-        guard !query.isEmpty else { return }
-
-        if !isSingleCharacterQuery, let forcedRemote = forcedRemoteSuggestionsForUITest() {
-            latestRemoteSuggestionQuery = query
-            latestRemoteSuggestions = forcedRemote
-            let merged = buildOmnibarSuggestions(
-                query: query,
-                engineName: searchConfiguration.displayName,
-                historyEntries: historyEntries,
-                openTabMatches: openTabMatches,
-                remoteQueries: forcedRemote,
-                resolvedURL: resolvedURL,
-                limit: 8
-            )
-            let forcedEffects = omnibarReduce(state: &omnibarState, event: .suggestionsUpdated(merged))
-            applyOmnibarEffects(forcedEffects)
-            refreshInlineCompletion()
-#if DEBUG
-            cmuxDebugLog(
-                "browser.omnibar.suggestions refresh=forcedRemote " +
-                "panel=\(panel.id.uuidString.prefix(5)) queryLen=\(query.utf8.count) items=\(merged.count)"
-            )
-#endif
-            return
-        }
-
-        guard remoteSuggestionsEnabled else { return }
-        guard !isSingleCharacterQuery else { return }
-        guard omnibarInputIntent(for: query) != .urlLike else { return }
-
-        // Keep current remote rows visible while fetching fresh predictions.
-        guard let engine = remoteSuggestionsEngine else { return }
-        isLoadingRemoteSuggestions = true
-        suggestionTask = Task {
-            let remote = await BrowserSearchSuggestionService.shared.suggestions(engine: engine, query: query)
-            if Task.isCancelled { return }
-
-            await MainActor.run {
-                guard !Task.isCancelled else { return }
-                guard addressBarFocused, !omnibarHasMarkedText else { return }
-                let current = omnibarState.buffer.trimmingCharacters(in: .whitespacesAndNewlines)
-                guard current == query else { return }
-                latestRemoteSuggestionQuery = query
-                latestRemoteSuggestions = remote
-                let merged = buildOmnibarSuggestions(
-                    query: query,
-                    engineName: searchConfiguration.displayName,
-                    historyEntries: panel.historyStore.suggestions(for: query, limit: 12),
-                    openTabMatches: matchingOpenTabSuggestions(for: query, limit: 12),
-                    remoteQueries: remote,
-                    resolvedURL: panel.resolveNavigableURL(from: query),
-                    limit: 8
-                )
-                let effects = omnibarReduce(state: &omnibarState, event: .suggestionsUpdated(merged))
-                applyOmnibarEffects(effects)
-                refreshInlineCompletion()
-                isLoadingRemoteSuggestions = false
-#if DEBUG
-                cmuxDebugLog(
-                    "browser.omnibar.suggestions refresh=remote " +
-                    "panel=\(panel.id.uuidString.prefix(5)) queryLen=\(query.utf8.count) " +
-                    "remote=\(remote.count) items=\(merged.count)"
-                )
-#endif
-            }
-        }
-    }
-
-    private func staleRemoteSuggestionsForDisplay(
-        query: String,
-        allowsRemoteSuggestions: Bool = true
-    ) -> [String] {
-        staleOmnibarRemoteSuggestionsForDisplay(
-            query: query,
-            previousRemoteQuery: latestRemoteSuggestionQuery,
-            previousRemoteSuggestions: latestRemoteSuggestions,
-            allowsRemoteSuggestions: allowsRemoteSuggestions
-        )
-    }
-
-    private func matchingOpenTabSuggestions(for query: String, limit: Int) -> [OmnibarOpenTabMatch] {
-        guard !query.isEmpty, limit > 0 else { return [] }
-        let singleCharacterQuery = omnibarSingleCharacterQuery(for: query)
-        let includeCurrentPanelForSingleCharacterQuery = singleCharacterQuery != nil
-        let currentPanelSnapshot = BrowserOpenTabSuggestionSnapshot(
-            workspaceId: panel.workspaceId,
-            panelId: panel.id,
-            url: panel.preferredURLStringForOmnibar(),
-            title: panel.pageTitle
-        )
-        let tabManager = AppDelegate.shared?.tabManagerFor(tabId: panel.workspaceId) ?? AppDelegate.shared?.tabManager
-        return tabManager?.matchingOpenBrowserTabSuggestions(
-            for: query,
-            currentWorkspaceId: panel.workspaceId,
-            currentPanelId: panel.id,
-            currentPanelSnapshot: currentPanelSnapshot,
-            includeCurrentPanelForSingleCharacterQuery: includeCurrentPanelForSingleCharacterQuery,
-            limit: limit
-        ) ?? []
-    }
-
-    private func forcedRemoteSuggestionsForUITest() -> [String]? {
-        let raw = ProcessInfo.processInfo.environment["CMUX_UI_TEST_REMOTE_SUGGESTIONS_JSON"]
-            ?? UserDefaults.standard.string(forKey: "CMUX_UI_TEST_REMOTE_SUGGESTIONS_JSON")
-        guard let raw,
-              let data = raw.data(using: .utf8),
-              let parsed = try? JSONSerialization.jsonObject(with: data) as? [Any] else {
-            return nil
-        }
-
-        let values = parsed.compactMap { item -> String? in
-            guard let s = item as? String else { return nil }
-            let trimmed = s.trimmingCharacters(in: .whitespacesAndNewlines)
-            return trimmed.isEmpty ? nil : trimmed
-        }
-        return values.isEmpty ? nil : values
-    }
-
-    private func applyOmnibarEffects(_ effects: OmnibarEffects) {
-        if effects.shouldCancelPendingSuggestionRefresh {
-            cancelPendingOmnibarSuggestionWork()
-        }
-        if effects.shouldClearInlineCompletion {
-            inlineCompletion = nil
-        }
-        if effects.shouldRefreshSuggestions {
-            omnibarSuggestionRefreshScheduler.scheduleRefresh()
-        }
-        if effects.shouldSelectAll {
-            omnibarSelectAllRequestId &+= 1
-        }
-        if effects.shouldBlurToWebView {
-            hideSuggestions()
-            // This transition is stateful: drop omnibar focus suppression before
-            // attempting responder handoff so WKWebView can actually become first responder.
-            panel.endSuppressWebViewFocusForAddressBar()
-            syncWebViewResponderPolicyWithViewState(reason: "effects.blurToWebView.preHandoff")
-            setAddressBarFocused(false, reason: "effects.blurToWebView")
-            DispatchQueue.main.async {
-                guard let window = panel.webView.window,
-                      !panel.webView.isHiddenOrHasHiddenAncestor else { return }
-                guard shouldApplyAddressBarExitFallback(in: window) else {
-#if DEBUG
-                    cmuxDebugLog(
-                        "browser.focus.addressBar.exit.handoff panel=\(panel.id.uuidString.prefix(5)) " +
-                        "result=skip_not_focused"
-                    )
-#endif
-                    NotificationCenter.default.post(name: .browserDidExitAddressBar, object: panel.id)
-                    return
-                }
-                syncWebViewResponderPolicyWithViewState(reason: "effects.blurToWebView.handoff")
-                panel.clearWebViewFocusSuppression()
-                let focusedWebView = window.makeFirstResponder(panel.webView)
-                if focusedWebView {
-                    panel.noteWebViewFocused()
-                }
-#if DEBUG
-                cmuxDebugLog(
-                    "browser.focus.addressBar.exit.handoff panel=\(panel.id.uuidString.prefix(5)) " +
-                    "focusedWebView=\(focusedWebView ? 1 : 0)"
-                )
-#endif
-                panel.restoreAddressBarPageFocusIfNeeded { restored in
-                    guard shouldApplyAddressBarExitFallback(in: window) else {
-#if DEBUG
-                        cmuxDebugLog(
-                            "browser.focus.addressBar.exit.handoff panel=\(panel.id.uuidString.prefix(5)) " +
-                            "result=skip_stale_restore restored=\(restored ? 1 : 0)"
-                        )
-#endif
-                        NotificationCenter.default.post(name: .browserDidExitAddressBar, object: panel.id)
-                        return
-                    }
-                    var hasWebViewResponder =
-                        browserFocusResponderChainContains(window.firstResponder, target: panel.webView)
-                    if !hasWebViewResponder {
-                        let fallbackFocusedWebView = window.makeFirstResponder(panel.webView)
-                        hasWebViewResponder = fallbackFocusedWebView
-#if DEBUG
-                        cmuxDebugLog(
-                            "browser.focus.addressBar.exit.handoff panel=\(panel.id.uuidString.prefix(5)) " +
-                            "fallbackFocusedWebView=\(fallbackFocusedWebView ? 1 : 0) " +
-                            "restored=\(restored ? 1 : 0)"
-                        )
-#endif
-                    }
-                    if hasWebViewResponder {
-                        panel.noteWebViewFocused()
-                    }
-                    NotificationCenter.default.post(name: .browserDidExitAddressBar, object: panel.id)
-                }
-            }
-        }
-    }
 }
 
-enum OmnibarInputIntent: Equatable {
-    case urlLike
-    case queryLike
-    case ambiguous
-}
 
-func omnibarInputIntent(for query: String) -> OmnibarInputIntent {
-    let trimmed = query.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return .ambiguous }
 
-    if resolveBrowserNavigableURL(trimmed) != nil {
-        return .urlLike
-    }
-
-    if trimmed.contains(" ") {
-        return .queryLike
-    }
-
-    if trimmed.contains(".") {
-        return .ambiguous
-    }
-
-    return .queryLike
-}
-
-func omnibarSuggestionCompletion(for suggestion: OmnibarSuggestion) -> String? {
-    switch suggestion.kind {
-    case .navigate(let url):
-        return url
-    case .history(let url, _):
-        return url
-    case .switchToTab(_, _, let url, _):
-        return url
-    default:
-        return nil
-    }
-}
-
-func omnibarSuggestionTitle(for suggestion: OmnibarSuggestion) -> String? {
-    switch suggestion.kind {
-    case .history(_, let title):
-        return title
-    case .switchToTab(_, _, _, let title):
-        return title
-    default:
-        return nil
-    }
-}
-
-func omnibarSuggestionMatchesTypedPrefix(
-    typedText: String,
-    suggestionCompletion: String,
-    suggestionTitle: String? = nil
-) -> Bool {
-    let trimmedQuery = typedText.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmedQuery.isEmpty else { return false }
-
-    let query = trimmedQuery.lowercased()
-    let trimmedCompletion = suggestionCompletion.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmedCompletion.isEmpty else { return false }
-    let loweredCompletion = trimmedCompletion.lowercased()
-
-    let schemeStripped = stripHTTPSchemePrefix(trimmedCompletion)
-    let schemeAndWWWStripped = stripHTTPSchemeAndWWWPrefix(trimmedCompletion)
-    let typedIncludesScheme = query.hasPrefix("https://") || query.hasPrefix("http://")
-    let typedIncludesWWWPrefix = query.hasPrefix("www.")
-
-    if typedIncludesScheme, loweredCompletion.hasPrefix(query) { return true }
-    if schemeStripped.hasPrefix(query) { return true }
-    if !typedIncludesWWWPrefix && schemeAndWWWStripped.hasPrefix(query) { return true }
-
-    let normalizedTitle = suggestionTitle?
-        .trimmingCharacters(in: .whitespacesAndNewlines)
-        .lowercased() ?? ""
-    if !normalizedTitle.isEmpty && normalizedTitle.hasPrefix(query) {
-        return true
-    }
-
-    return false
-}
-
-func omnibarSuggestionSupportsAutocompletion(query: String, suggestion: OmnibarSuggestion) -> Bool {
-    if case .search = suggestion.kind { return false }
-    if case .remote = suggestion.kind { return false }
-    guard let completion = omnibarSuggestionCompletion(for: suggestion) else { return false }
-    // Reject URLs whose host lacks a TLD (e.g. "https://news." → host "news").
-    if let components = URLComponents(string: completion),
-       let host = components.host?.lowercased() {
-        let trimmedHost = host.hasSuffix(".") ? String(host.dropLast()) : host
-        if !trimmedHost.contains(".") { return false }
-    }
-    let title = omnibarSuggestionTitle(for: suggestion)
-    return omnibarSuggestionMatchesTypedPrefix(
-        typedText: query,
-        suggestionCompletion: completion,
-        suggestionTitle: title
-    )
-}
-
-func omnibarSingleCharacterQuery(for query: String) -> String? {
-    let trimmed = query.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-    guard trimmed.utf16.count == 1 else { return nil }
-    return trimmed
-}
-
-func omnibarStrippedURL(_ value: String) -> String {
-    return stripHTTPSchemeAndWWWPrefix(value)
-}
-
-func omnibarScoringCandidate(_ value: String) -> String {
-    let trimmed = value.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return "" }
-
-    if let components = URLComponents(string: trimmed), let host = components.host?.lowercased() {
-        let hostWithoutWWW = host.hasPrefix("www.") ? String(host.dropFirst(4)) : host
-        let normalizedScheme = components.scheme?.lowercased()
-        let isDefaultPort = (normalizedScheme == "http" && components.port == 80)
-            || (normalizedScheme == "https" && components.port == 443)
-        let portSuffix = {
-            guard let port = components.port, !isDefaultPort else { return "" }
-            return ":\(port)"
-        }()
-
-        var normalized = "\(hostWithoutWWW)\(portSuffix)"
-        let path = components.percentEncodedPath
-        if !path.isEmpty && path != "/" {
-            normalized += path
-        } else if path == "/" {
-            normalized += "/"
-        }
-
-        if let query = components.percentEncodedQuery, !query.isEmpty {
-            normalized += "?\(query)"
-        }
-        if let fragment = components.percentEncodedFragment, !fragment.isEmpty {
-            normalized += "#\(fragment)"
-        }
-        return normalized
-    }
-
-    return stripHTTPSchemeAndWWWPrefix(trimmed)
-}
-
-func omnibarHasSingleCharacterPrefixMatch(query: String, url: String, title: String?) -> Bool {
-    guard let trimmedQuery = omnibarSingleCharacterQuery(for: query) else { return false }
-
-    let normalizedURL = omnibarStrippedURL(url).lowercased()
-    let normalizedTitle = title?.trimmingCharacters(in: .whitespacesAndNewlines).lowercased() ?? ""
-    return normalizedURL.hasPrefix(trimmedQuery) || normalizedTitle.hasPrefix(trimmedQuery)
-}
-
-func buildOmnibarSuggestions(
-    query: String,
-    engineName: String,
-    historyEntries: [BrowserHistoryStore.Entry],
-    openTabMatches: [OmnibarOpenTabMatch] = [],
-    remoteQueries: [String],
-    resolvedURL: URL?,
-    limit: Int = 8,
-    now: Date = Date()
-) -> [OmnibarSuggestion] {
-    guard limit > 0 else { return [] }
-
-    let trimmedQuery = query.trimmingCharacters(in: .whitespacesAndNewlines)
-    if trimmedQuery.isEmpty {
-        return Array(historyEntries.prefix(limit).map { .history($0) })
-    }
-    let singleCharacterQuery = omnibarSingleCharacterQuery(for: trimmedQuery)
-    let isSingleCharacterQuery = singleCharacterQuery != nil
-    let shouldIncludeRemoteSuggestions = !isSingleCharacterQuery
-    let filteredHistoryEntries: [BrowserHistoryStore.Entry]
-    let filteredOpenTabMatches: [OmnibarOpenTabMatch]
-    if let singleCharacterQuery {
-        filteredHistoryEntries = historyEntries.filter {
-            omnibarHasSingleCharacterPrefixMatch(query: singleCharacterQuery, url: $0.url, title: $0.title)
-        }
-        filteredOpenTabMatches = openTabMatches.filter {
-            omnibarHasSingleCharacterPrefixMatch(query: singleCharacterQuery, url: $0.url, title: $0.title)
-        }
-    } else {
-        filteredHistoryEntries = historyEntries
-        filteredOpenTabMatches = openTabMatches
-    }
-
-    let shouldSuppressSingleCharacterSearchResult = isSingleCharacterQuery
-        && (!filteredHistoryEntries.isEmpty || !filteredOpenTabMatches.isEmpty)
-
-    struct RankedSuggestion {
-        let suggestion: OmnibarSuggestion
-        let score: Double
-        let order: Int
-        let isAutocompletableMatch: Bool
-        let kindPriority: Int
-    }
-
-    var bestByCompletion: [String: RankedSuggestion] = [:]
-    var order = 0
-    let intent = omnibarInputIntent(for: trimmedQuery)
-    let normalizedQuery = trimmedQuery.lowercased()
-
-    func suggestionPriority(for kind: OmnibarSuggestion.Kind) -> Int {
-        switch kind {
-        case .search:
-            return 300
-        case .remote:
-            return 350
-        default:
-            return 0
-        }
-    }
-
-    func completionScore(for candidate: String) -> Double {
-        let c = candidate.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-        let q = normalizedQuery
-        guard !c.isEmpty, !q.isEmpty else { return 0 }
-
-        let scoringCandidate = omnibarScoringCandidate(c)
-        if !scoringCandidate.isEmpty {
-            if scoringCandidate == q { return 260 }
-            if scoringCandidate.hasPrefix(q) { return 220 }
-            if scoringCandidate.contains(q) { return 150 }
-        }
-
-        if c == q { return 240 }
-        if c.hasPrefix(q) { return 170 }
-        if c.contains(q) { return 95 }
-        return 0
-    }
-
-    func insert(_ suggestion: OmnibarSuggestion, score: Double) {
-        let key = suggestion.completion.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-        guard !key.isEmpty else { return }
-        let isAutocompletableMatch = omnibarSuggestionSupportsAutocompletion(query: trimmedQuery, suggestion: suggestion)
-
-        let ranked = RankedSuggestion(
-            suggestion: suggestion,
-            score: score,
-            order: order,
-            isAutocompletableMatch: isAutocompletableMatch,
-            kindPriority: suggestionPriority(for: suggestion.kind)
-        )
-        order += 1
-        if let existing = bestByCompletion[key] {
-            let shouldReplaceExisting: Bool = {
-                // For identical completions, keep "go to URL" over "switch to tab" so
-                // pressing Enter performs navigation unless the user explicitly picks a tab row.
-                switch (existing.suggestion.kind, ranked.suggestion.kind) {
-                case (.navigate, .switchToTab):
-                    return false
-                case (.switchToTab, .navigate):
-                    return true
-                default:
-                    return ranked.score > existing.score
-                }
-            }()
-            if shouldReplaceExisting {
-                bestByCompletion[key] = ranked
-            }
-        } else {
-            bestByCompletion[key] = ranked
-        }
-    }
-
-    if !(isSingleCharacterQuery && shouldSuppressSingleCharacterSearchResult) {
-        let searchBaseScore: Double
-        switch intent {
-        case .queryLike: searchBaseScore = 820
-        case .ambiguous: searchBaseScore = 540
-        case .urlLike: searchBaseScore = 140
-        }
-        insert(.search(engineName: engineName, query: trimmedQuery), score: searchBaseScore + completionScore(for: trimmedQuery))
-    }
-
-    if let resolvedURL {
-        let completion = resolvedURL.absoluteString
-        let navigateBaseScore: Double
-        switch intent {
-        case .urlLike: navigateBaseScore = 1_020
-        case .ambiguous: navigateBaseScore = 760
-        case .queryLike: navigateBaseScore = 470
-        }
-        insert(.navigate(url: completion), score: navigateBaseScore + completionScore(for: completion))
-    }
-
-    for (index, entry) in filteredHistoryEntries.prefix(max(limit * 2, limit)).enumerated() {
-        let intentBaseScore: Double
-        switch intent {
-        case .urlLike: intentBaseScore = 780
-        case .ambiguous: intentBaseScore = 690
-        case .queryLike: intentBaseScore = 600
-        }
-        let urlMatch = completionScore(for: entry.url)
-        let titleMatch = completionScore(for: entry.title ?? "") * 0.6
-        let ageHours = max(0, now.timeIntervalSince(entry.lastVisited) / 3600)
-        let recencyScore = max(0, 75 - (ageHours / 5))
-        let visitScore = min(95, log1p(Double(max(1, entry.visitCount))) * 32)
-        let typedScore = min(230, log1p(Double(max(0, entry.typedCount))) * 100)
-        let typedRecencyScore: Double
-        if let lastTypedAt = entry.lastTypedAt {
-            let typedAgeHours = max(0, now.timeIntervalSince(lastTypedAt) / 3600)
-            typedRecencyScore = max(0, 80 - (typedAgeHours / 5))
-        } else {
-            typedRecencyScore = 0
-        }
-        let positionScore = Double(max(0, 16 - index))
-        let total = intentBaseScore + urlMatch + titleMatch + recencyScore + visitScore + typedScore + typedRecencyScore + positionScore
-        insert(.history(entry), score: total)
-    }
-
-    for (index, match) in filteredOpenTabMatches.prefix(limit).enumerated() {
-        let intentBaseScore: Double
-        switch intent {
-        case .urlLike: intentBaseScore = 1_180
-        case .ambiguous: intentBaseScore = 980
-        case .queryLike: intentBaseScore = 820
-        }
-        let urlMatch = completionScore(for: match.url)
-        let titleMatch = completionScore(for: match.title ?? "") * 0.65
-        let positionScore = Double(max(0, 14 - index)) * 0.9
-        let resolvedURLBonus: Double
-        if let resolvedURL,
-           resolvedURL.absoluteString.caseInsensitiveCompare(match.url) == .orderedSame {
-            resolvedURLBonus = 120
-        } else {
-            resolvedURLBonus = 0
-        }
-        let total = intentBaseScore + urlMatch + titleMatch + positionScore + resolvedURLBonus
-        if match.isKnownOpenTab {
-            insert(
-                .switchToTab(tabId: match.tabId, panelId: match.panelId, url: match.url, title: match.title),
-                score: total
-            )
-        } else {
-            insert(
-                OmnibarSuggestion.history(url: match.url, title: match.title),
-                score: total
-            )
-        }
-    }
-
-    if shouldIncludeRemoteSuggestions {
-        for (index, remoteQuery) in remoteQueries.prefix(limit).enumerated() {
-            let trimmedRemote = remoteQuery.trimmingCharacters(in: .whitespacesAndNewlines)
-            guard !trimmedRemote.isEmpty else { continue }
-
-            let remoteBaseScore: Double
-            switch intent {
-            case .queryLike: remoteBaseScore = 690
-            case .ambiguous: remoteBaseScore = 450
-            case .urlLike: remoteBaseScore = 110
-            }
-            let positionScore = Double(max(0, 14 - index)) * 0.9
-            let total = remoteBaseScore + completionScore(for: trimmedRemote) + positionScore
-            insert(.remoteSearchSuggestion(trimmedRemote), score: total)
-        }
-    }
-
-    let sorted = bestByCompletion.values.sorted { lhs, rhs in
-        if lhs.isAutocompletableMatch != rhs.isAutocompletableMatch {
-            return lhs.isAutocompletableMatch
-        }
-        if lhs.score != rhs.score { return lhs.score > rhs.score }
-        if lhs.kindPriority != rhs.kindPriority {
-            return lhs.kindPriority < rhs.kindPriority
-        }
-        if lhs.order != rhs.order { return lhs.order < rhs.order }
-        return lhs.suggestion.completion < rhs.suggestion.completion
-    }
-    let suggestions = Array(sorted.map(\.suggestion).prefix(limit))
-    return prioritizedAutocompletionSuggestions(suggestions: Array(suggestions), for: trimmedQuery)
-}
-
-private func prioritizedAutocompletionSuggestions(suggestions: [OmnibarSuggestion], for query: String) -> [OmnibarSuggestion] {
-    guard let preferred = omnibarPreferredAutocompletionSuggestionIndex(
-        suggestions: suggestions,
-        query: query
-    ) else {
-        return suggestions
-    }
-
-    guard preferred != 0 else { return suggestions }
-
-    var reordered = suggestions
-    let suggestion = reordered.remove(at: preferred)
-    reordered.insert(suggestion, at: 0)
-    return reordered
-}
-
-private func omnibarPreferredAutocompletionSuggestionIndex(
-    suggestions: [OmnibarSuggestion],
-    query: String
-) -> Int? {
-    guard !query.isEmpty else { return nil }
-
-    var candidates: [(idx: Int, suffixLength: Int)] = []
-    for (idx, suggestion) in suggestions.enumerated() {
-        guard omnibarSuggestionSupportsAutocompletion(query: query, suggestion: suggestion) else { continue }
-        guard let completion = omnibarSuggestionCompletion(for: suggestion) else { continue }
-        let displayCompletion = omnibarSuggestionMatchesTypedPrefix(
-            typedText: query,
-            suggestionCompletion: completion,
-            suggestionTitle: omnibarSuggestionTitle(for: suggestion)
-        ) ? completion : ""
-        guard !displayCompletion.isEmpty else { continue }
-
-        let suffixLength = max(
-            0,
-            omnibarSuggestionDisplayText(forPrefixing: displayCompletion, query: query).utf16.count - query.utf16.count
-        )
-        candidates.append((idx: idx, suffixLength: suffixLength))
-    }
-
-    guard let preferred = candidates.min(by: {
-        if $0.suffixLength != $1.suffixLength {
-            return $0.suffixLength < $1.suffixLength
-        }
-        return $0.idx < $1.idx
-    })?.idx else {
-        return nil
-    }
-
-    return preferred
-}
-
-private func omnibarSuggestionDisplayText(forPrefixing completion: String, query: String) -> String {
-    let typedIncludesScheme = query.hasPrefix("https://") || query.hasPrefix("http://")
-    let typedIncludesWWWPrefix = query.hasPrefix("www.")
-    if typedIncludesScheme {
-        return completion
-    }
-    if typedIncludesWWWPrefix {
-        return stripHTTPSchemePrefix(completion)
-    }
-    return stripHTTPSchemeAndWWWPrefix(completion)
-}
-
-func staleOmnibarRemoteSuggestionsForDisplay(
-    query: String,
-    previousRemoteQuery: String,
-    previousRemoteSuggestions: [String],
-    allowsRemoteSuggestions: Bool = true,
-    limit: Int = 8
-) -> [String] {
-    guard allowsRemoteSuggestions else { return [] }
-    let trimmedQuery = query.trimmingCharacters(in: .whitespacesAndNewlines)
-    let trimmedPreviousQuery = previousRemoteQuery.trimmingCharacters(in: .whitespacesAndNewlines)
-    let loweredQuery = trimmedQuery.lowercased()
-    let loweredPreviousQuery = trimmedPreviousQuery.lowercased()
-    guard !trimmedQuery.isEmpty, !trimmedPreviousQuery.isEmpty else { return [] }
-    guard loweredQuery == loweredPreviousQuery || loweredQuery.hasPrefix(loweredPreviousQuery) || loweredPreviousQuery.hasPrefix(loweredQuery) else {
-        return []
-    }
-    guard !previousRemoteSuggestions.isEmpty else { return [] }
-    let sanitized = previousRemoteSuggestions.compactMap { raw -> String? in
-        let trimmed = raw.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty else { return nil }
-        return trimmed
-    }
-
-    if sanitized.isEmpty {
-        return []
-    }
-    return Array(sanitized.prefix(limit))
-}
-
-func omnibarInlineCompletionForDisplay(
-    typedText: String,
-    suggestions: [OmnibarSuggestion],
-    isFocused: Bool,
-    selectionRange: NSRange,
-    hasMarkedText: Bool
-) -> OmnibarInlineCompletion? {
-    guard isFocused else { return nil }
-    guard !hasMarkedText else { return nil }
-
-    let query = typedText.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !query.isEmpty else { return nil }
-    let loweredQuery = query.lowercased()
-    let typedIncludesScheme = loweredQuery.hasPrefix("https://") || loweredQuery.hasPrefix("http://")
-    let typedIncludesWWWPrefix = loweredQuery.hasPrefix("www.")
-    let queryCount = query.utf16.count
-
-    let urlCandidate = suggestions.first { suggestion in
-        guard let completion = omnibarSuggestionCompletion(for: suggestion) else { return false }
-        return omnibarSuggestionMatchesTypedPrefix(
-            typedText: query,
-            suggestionCompletion: completion,
-            suggestionTitle: omnibarSuggestionTitle(for: suggestion)
-        )
-    }
-    guard let candidate = urlCandidate else {
-        return nil
-    }
-
-    let acceptedText = candidate.completion
-    let displayText: String
-    if typedQueryHasExplicitPathOrQuery(query) {
-        if typedIncludesScheme {
-            displayText = acceptedText
-        } else if typedIncludesWWWPrefix {
-            displayText = stripHTTPSchemePrefix(acceptedText)
-        } else {
-            displayText = stripHTTPSchemeAndWWWPrefix(acceptedText)
-        }
-    } else if let hostOnlyDisplay = inlineCompletionHostDisplayText(
-        for: acceptedText,
-        typedIncludesScheme: typedIncludesScheme,
-        typedIncludesWWWPrefix: typedIncludesWWWPrefix
-    ) {
-        displayText = hostOnlyDisplay
-    } else {
-        if typedIncludesScheme {
-            displayText = acceptedText
-        } else if typedIncludesWWWPrefix {
-            displayText = stripHTTPSchemePrefix(acceptedText)
-        } else {
-            displayText = stripHTTPSchemeAndWWWPrefix(acceptedText)
-        }
-    }
-
-    guard omnibarSuggestionSupportsAutocompletion(query: query, suggestion: candidate) else { return nil }
-    // The display text must start with the typed query so the inline completion
-    // visually extends what the user typed rather than replacing it (e.g. a
-    // history entry matched via title "localhost:3000" whose URL is google.com
-    // should not replace a typed "l" with "g").
-    guard displayText.lowercased().hasPrefix(loweredQuery) else { return nil }
-    guard displayText.utf16.count > queryCount else {
-        return nil
-    }
-
-    let displayCount = displayText.utf16.count
-
-    let resolvedSelectionRange: NSRange = {
-        if selectionRange.location == NSNotFound {
-            return NSRange(location: queryCount, length: 0)
-        }
-        let clampedLocation = min(selectionRange.location, displayCount)
-        let remaining = max(0, displayCount - clampedLocation)
-        let clampedLength = min(selectionRange.length, remaining)
-        return NSRange(location: clampedLocation, length: clampedLength)
-    }()
-
-    let suffixRange = NSRange(location: queryCount, length: max(0, displayCount - queryCount))
-    let isCaretAtTypedBoundary = (resolvedSelectionRange.length == 0 && resolvedSelectionRange.location == queryCount)
-    let isSuffixSelection = NSEqualRanges(resolvedSelectionRange, suffixRange)
-    let isSelectAllSelection = (resolvedSelectionRange.location == 0 && resolvedSelectionRange.length == displayCount)
-    // Command+A can briefly report just the typed prefix selection before the full
-    // select-all range lands. Keep inline completion alive through that transition.
-    let typedPrefixSelection = NSRange(location: 0, length: queryCount)
-    let isTypedPrefixSelection = NSEqualRanges(resolvedSelectionRange, typedPrefixSelection)
-    guard isCaretAtTypedBoundary || isSuffixSelection || isSelectAllSelection || isTypedPrefixSelection else {
-        return nil
-    }
-
-    return OmnibarInlineCompletion(typedText: query, displayText: displayText, acceptedText: acceptedText)
-}
-
-func omnibarDesiredSelectionRangeForInlineCompletion(
-    currentSelection: NSRange,
-    inlineCompletion: OmnibarInlineCompletion
-) -> NSRange {
-    let typedCount = inlineCompletion.typedText.utf16.count
-    let typedPrefixSelection = NSRange(location: 0, length: typedCount)
-    let displayCount = inlineCompletion.displayText.utf16.count
-    let isSelectAll = currentSelection.location == 0 && currentSelection.length == displayCount
-    if isSelectAll ||
-        NSEqualRanges(currentSelection, inlineCompletion.suffixRange) ||
-        NSEqualRanges(currentSelection, typedPrefixSelection) {
-        return currentSelection
-    }
-    return inlineCompletion.suffixRange
-}
-
-func omnibarPublishedBufferTextForFieldChange(
-    fieldValue: String,
-    inlineCompletion: OmnibarInlineCompletion?,
-    selectionRange: NSRange?,
-    hasMarkedText: Bool
-) -> String {
-    guard !hasMarkedText else { return fieldValue }
-    guard let inlineCompletion else { return fieldValue }
-    guard fieldValue == inlineCompletion.displayText else { return fieldValue }
-    guard let selectionRange else { return inlineCompletion.typedText }
-
-    let typedCount = inlineCompletion.typedText.utf16.count
-    let displayCount = inlineCompletion.displayText.utf16.count
-    let typedPrefixSelection = NSRange(location: 0, length: typedCount)
-    let isCaretAtTypedBoundary = selectionRange.location == typedCount && selectionRange.length == 0
-    let isSuffixSelection = NSEqualRanges(selectionRange, inlineCompletion.suffixRange)
-    let isSelectAllSelection = selectionRange.location == 0 && selectionRange.length == displayCount
-    let isTypedPrefixSelection = NSEqualRanges(selectionRange, typedPrefixSelection)
-    if isCaretAtTypedBoundary || isSuffixSelection || isSelectAllSelection || isTypedPrefixSelection {
-        return inlineCompletion.typedText
-    }
-
-    return fieldValue
-}
-
-func omnibarInlineCompletionIfBufferMatchesTypedPrefix(
-    bufferText: String,
-    inlineCompletion: OmnibarInlineCompletion?
-) -> OmnibarInlineCompletion? {
-    guard let inlineCompletion else { return nil }
-    guard bufferText == inlineCompletion.typedText else { return nil }
-    return inlineCompletion
-}
-
-func omnibarPrefixAfterDeletingTrailingWord(from text: String) -> String {
-    let nsText = text as NSString
-    let fullRange = NSRange(location: 0, length: nsText.length)
-    var deletionStart = nsText.length
-    nsText.enumerateSubstrings(in: fullRange, options: [.byWords, .reverse]) { _, range, _, stop in
-        deletionStart = range.location
-        stop.pointee = true
-    }
-    return nsText.substring(to: deletionStart)
-}
-
-private func typedQueryHasExplicitPathOrQuery(_ typedQuery: String) -> Bool {
-    var normalized = typedQuery.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-    if normalized.hasPrefix("https://") {
-        normalized.removeFirst("https://".count)
-    } else if normalized.hasPrefix("http://") {
-        normalized.removeFirst("http://".count)
-    }
-    return normalized.contains("/") || normalized.contains("?") || normalized.contains("#")
-}
-
-private func inlineCompletionHostDisplayText(
-    for acceptedText: String,
-    typedIncludesScheme: Bool,
-    typedIncludesWWWPrefix: Bool
-) -> String? {
-    guard let components = URLComponents(string: acceptedText),
-          var host = components.host?.trimmingCharacters(in: .whitespacesAndNewlines).lowercased(),
-          !host.isEmpty else {
-        return nil
-    }
-
-    if !typedIncludesWWWPrefix, host.hasPrefix("www.") {
-        host.removeFirst("www.".count)
-    }
-
-    let portSuffix: String
-    if let port = components.port {
-        let scheme = components.scheme?.lowercased()
-        let isDefaultPort =
-            (scheme == "https" && port == 443) ||
-            (scheme == "http" && port == 80)
-        portSuffix = isDefaultPort ? "" : ":\(port)"
-    } else {
-        portSuffix = ""
-    }
-
-    let hostWithPort = "\(host)\(portSuffix)"
-    if typedIncludesScheme {
-        let scheme = (components.scheme?.lowercased() == "http") ? "http" : "https"
-        return "\(scheme)://\(hostWithPort)"
-    }
-    return hostWithPort
-}
-
-private func stripHTTPSchemePrefix(_ raw: String) -> String {
-    var normalized = raw.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-    if normalized.hasPrefix("https://") {
-        normalized.removeFirst("https://".count)
-    } else if normalized.hasPrefix("http://") {
-        normalized.removeFirst("http://".count)
-    }
-    return normalized
-}
-
-private func stripHTTPSchemeAndWWWPrefix(_ raw: String) -> String {
-    var normalized = stripHTTPSchemePrefix(raw)
-    if normalized.hasPrefix("www.") {
-        normalized.removeFirst("www.".count)
-    }
-    return normalized
-}
-
-private struct OmnibarPillFramePreferenceKey: PreferenceKey {
-    static var defaultValue: CGRect = .zero
-
-    static func reduce(value: inout CGRect, nextValue: () -> CGRect) {
-        let next = nextValue()
-        if next != .zero {
-            value = next
-        }
-    }
-}
-
-private struct BrowserAddressBarHeightPreferenceKey: PreferenceKey {
-    static var defaultValue: CGFloat = 0
-
-    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-        value = max(value, nextValue())
-    }
-}
-
-private struct BrowserAddressBarWidthPreferenceKey: PreferenceKey {
-    static var defaultValue: CGFloat = 0
-
-    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-        value = max(value, nextValue())
-    }
-}
-
-// MARK: - Omnibar State Machine
-
-struct OmnibarState: Equatable {
-    var isFocused: Bool = false
-    var currentURLString: String = ""
-    var buffer: String = ""
-    var suggestions: [OmnibarSuggestion] = []
-    var selectedSuggestionIndex: Int = 0
-    var selectedSuggestionID: String?
-    /// True only while the current suggestion selection came from an explicit
-    /// user action (arrow keys, Ctrl+N/P). Automatic highlighting (preferred
-    /// autocompletion pick, popup reopen, pointer hover) leaves this false so
-    /// a row auto-selected for an older query can never hijack Return.
-    var selectionIsExplicit: Bool = false
-    var isUserEditing: Bool = false
-}
-
-enum OmnibarEvent: Equatable {
-    case focusGained(currentURLString: String, shouldSelectAll: Bool = false)
-    case focusReasserted(shouldSelectAll: Bool = true)
-    case focusLostRevertBuffer(currentURLString: String)
-    case focusLostPreserveBuffer(currentURLString: String)
-    case panelURLChanged(currentURLString: String)
-    case bufferChanged(String)
-    case suggestionsUpdated([OmnibarSuggestion])
-    case moveSelection(delta: Int)
-    case highlightIndex(Int)
-    case escape
-}
-
-struct OmnibarEffects: Equatable {
-    var shouldSelectAll: Bool = false
-    var shouldBlurToWebView: Bool = false
-    var shouldRefreshSuggestions: Bool = false
-    var shouldClearInlineCompletion: Bool = false
-    var shouldCancelPendingSuggestionRefresh: Bool = false
-}
-
-@discardableResult
-func omnibarReduce(state: inout OmnibarState, event: OmnibarEvent) -> OmnibarEffects {
-    var effects = OmnibarEffects()
-
-    switch event {
-    case .focusGained(let url, let shouldSelectAll):
-        state.isFocused = true
-        state.currentURLString = url
-        state.buffer = url
-        state.isUserEditing = false
-        state.suggestions = []
-        state.selectedSuggestionIndex = 0
-        state.selectedSuggestionID = nil
-        state.selectionIsExplicit = false
-        effects.shouldSelectAll = shouldSelectAll
-        effects.shouldCancelPendingSuggestionRefresh = true
-
-    case .focusReasserted(let shouldSelectAll):
-        state.isFocused = true
-        effects.shouldSelectAll = shouldSelectAll
-        if shouldSelectAll {
-            // A Cmd+L style reassert restarts editing from the full selected
-            // text; an earlier arrow selection no longer reflects Return
-            // intent. Plain focus restoration (no select-all) keeps it.
-            state.selectionIsExplicit = false
-        }
-
-    case .focusLostRevertBuffer(let url):
-        state.isFocused = false
-        state.currentURLString = url
-        state.buffer = url
-        state.isUserEditing = false
-        state.suggestions = []
-        state.selectedSuggestionIndex = 0
-        state.selectedSuggestionID = nil
-        state.selectionIsExplicit = false
-        effects.shouldCancelPendingSuggestionRefresh = true
-
-    case .focusLostPreserveBuffer(let url):
-        state.isFocused = false
-        state.currentURLString = url
-        state.isUserEditing = false
-        state.suggestions = []
-        state.selectedSuggestionIndex = 0
-        state.selectedSuggestionID = nil
-        state.selectionIsExplicit = false
-        effects.shouldCancelPendingSuggestionRefresh = true
-
-    case .panelURLChanged(let url):
-        state.currentURLString = url
-        if !state.isUserEditing {
-            state.buffer = url
-            state.suggestions = []
-            state.selectedSuggestionIndex = 0
-            state.selectedSuggestionID = nil
-            state.selectionIsExplicit = false
-            effects.shouldCancelPendingSuggestionRefresh = true
-        }
-
-    case .bufferChanged(let newValue):
-        let bufferChanged = state.buffer != newValue
-        state.buffer = newValue
-        if state.isFocused {
-            state.isUserEditing = (newValue != state.currentURLString)
-            state.selectedSuggestionIndex = 0
-            state.selectedSuggestionID = nil
-            state.selectionIsExplicit = false
-            effects.shouldRefreshSuggestions = true
-            effects.shouldClearInlineCompletion = bufferChanged
-        }
-
-    case .suggestionsUpdated(let items):
-        let previousItems = state.suggestions
-        let previousSelectedID = state.selectedSuggestionID
-        state.suggestions = items
-        if items.isEmpty {
-            state.selectedSuggestionIndex = 0
-            state.selectedSuggestionID = nil
-            state.selectionIsExplicit = false
-        } else if let previousSelectedID,
-                  let existingIdx = items.firstIndex(where: { $0.id == previousSelectedID }) {
-            // Same row carried across a refresh: an explicit selection stays explicit.
-            state.selectedSuggestionIndex = existingIdx
-            state.selectedSuggestionID = items[existingIdx].id
-        } else if let preferredSuggestionIndex = omnibarPreferredAutocompletionSuggestionIndex(
-            suggestions: items,
-            query: state.buffer
-        ) {
-            state.selectedSuggestionIndex = preferredSuggestionIndex
-            state.selectedSuggestionID = items[preferredSuggestionIndex].id
-            state.selectionIsExplicit = false
-        } else if previousItems.isEmpty {
-            // Popup reopened: start keyboard focus from the first row.
-            state.selectedSuggestionIndex = 0
-            state.selectedSuggestionID = items[0].id
-            state.selectionIsExplicit = false
-        } else {
-            state.selectedSuggestionIndex = min(max(0, state.selectedSuggestionIndex), items.count - 1)
-            state.selectedSuggestionID = items[state.selectedSuggestionIndex].id
-            state.selectionIsExplicit = false
-        }
-
-    case .moveSelection(let delta):
-        guard !state.suggestions.isEmpty else { break }
-        state.selectedSuggestionIndex = min(
-            max(0, state.selectedSuggestionIndex + delta),
-            state.suggestions.count - 1
-        )
-        state.selectedSuggestionID = state.suggestions[state.selectedSuggestionIndex].id
-        state.selectionIsExplicit = true
-
-    case .highlightIndex(let idx):
-        guard !state.suggestions.isEmpty else { break }
-        state.selectedSuggestionIndex = min(max(0, idx), state.suggestions.count - 1)
-        state.selectedSuggestionID = state.suggestions[state.selectedSuggestionIndex].id
-        // Pointer hover tracks the highlight but is not an explicit selection:
-        // the popup can appear underneath a stationary cursor.
-        state.selectionIsExplicit = false
-
-    case .escape:
-        guard state.isFocused else { break }
-        // Chrome semantics:
-        // - If user input is in progress OR the popup is open: revert to the page URL and select-all.
-        // - Otherwise: exit omnibar focus.
-        if state.isUserEditing || !state.suggestions.isEmpty {
-            state.isUserEditing = false
-            state.buffer = state.currentURLString
-            state.suggestions = []
-            state.selectedSuggestionIndex = 0
-            state.selectedSuggestionID = nil
-            state.selectionIsExplicit = false
-            effects.shouldSelectAll = true
-            effects.shouldCancelPendingSuggestionRefresh = true
-        } else {
-            effects.shouldBlurToWebView = true
-        }
-    }
-
-    return effects
-}
-
-struct OmnibarSuggestion: Identifiable, Hashable {
-    enum Kind: Hashable {
-        case search(engineName: String, query: String)
-        case navigate(url: String)
-        case history(url: String, title: String?)
-        case switchToTab(tabId: UUID, panelId: UUID, url: String, title: String?)
-        case remote(query: String)
-    }
-
-    let kind: Kind
-
-    // Stable identity prevents row teardown/rebuild flicker while typing.
-    var id: String {
-        switch kind {
-        case .search(let engineName, let query):
-            return "search|\(engineName.lowercased())|\(query.lowercased())"
-        case .navigate(let url):
-            return "navigate|\(url.lowercased())"
-        case .history(let url, _):
-            return "history|\(url.lowercased())"
-        case .switchToTab(let tabId, let panelId, let url, _):
-            return "switch-tab|\(tabId.uuidString.lowercased())|\(panelId.uuidString.lowercased())|\(url.lowercased())"
-        case .remote(let query):
-            return "remote|\(query.lowercased())"
-        }
-    }
-
-    var completion: String {
-        switch kind {
-        case .search(_, let q): return q
-        case .navigate(let url): return url
-        case .history(let url, _): return url
-        case .switchToTab(_, _, let url, _): return url
-        case .remote(let q): return q
-        }
-    }
-
-    var primaryText: String {
-        switch kind {
-        case .search(let engineName, let q):
-            return "Search \(engineName) for \"\(q)\""
-        case .navigate(let url):
-            return Self.displayURLText(for: url)
-        case .history(let url, let title):
-            return (title?.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty == false)
-                ? Self.singleLineText(title) : Self.displayURLText(for: url)
-        case .switchToTab(_, _, let url, let title):
-            return (title?.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty == false)
-                ? Self.singleLineText(title) : Self.displayURLText(for: url)
-        case .remote(let q):
-            return q
-        }
-    }
-
-    var listText: String {
-        switch kind {
-        case .history(let url, let title), .switchToTab(_, _, let url, let title):
-            let titleOneline = Self.singleLineText(title)
-            guard !titleOneline.isEmpty else { return Self.displayURLText(for: url) }
-            return "\(titleOneline) — \(Self.displayURLText(for: url))"
-        default:
-            return primaryText
-        }
-    }
-
-    var secondaryText: String? {
-        switch kind {
-        case .history(let url, let title):
-            let titleOneline = Self.singleLineText(title)
-            return titleOneline.isEmpty ? nil : Self.displayURLText(for: url)
-        case .switchToTab(_, _, let url, let title):
-            let titleOneline = Self.singleLineText(title)
-            return titleOneline.isEmpty ? nil : Self.displayURLText(for: url)
-        default:
-            return nil
-        }
-    }
-
-    var trailingBadgeText: String? {
-        switch kind {
-        case .switchToTab:
-            return String(localized: "browser.switchToTab", defaultValue: "Switch to tab")
-        default:
-            return nil
-        }
-    }
-
-    var isHistoryRemovable: Bool {
-        if case .history = kind { return true }
-        return false
-    }
-
-    static func history(_ entry: BrowserHistoryStore.Entry) -> OmnibarSuggestion {
-        OmnibarSuggestion(kind: .history(url: entry.url, title: entry.title))
-    }
-
-    static func history(url: String, title: String?) -> OmnibarSuggestion {
-        OmnibarSuggestion(kind: .history(url: url, title: title))
-    }
-
-    static func search(engineName: String, query: String) -> OmnibarSuggestion {
-        OmnibarSuggestion(kind: .search(engineName: engineName, query: query))
-    }
-
-    static func navigate(url: String) -> OmnibarSuggestion {
-        OmnibarSuggestion(kind: .navigate(url: url))
-    }
-
-    static func switchToTab(tabId: UUID, panelId: UUID, url: String, title: String?) -> OmnibarSuggestion {
-        OmnibarSuggestion(kind: .switchToTab(tabId: tabId, panelId: panelId, url: url, title: title))
-    }
-
-    private static func singleLineText(_ value: String?) -> String {
-        var normalized = (value ?? "").replacingOccurrences(of: "\r", with: " ")
-            .replacingOccurrences(of: "\n", with: " ")
-            .replacingOccurrences(of: "\t", with: " ")
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-        while normalized.contains("  ") {
-            let collapsed = normalized.replacingOccurrences(of: "  ", with: " ")
-            if collapsed == normalized { break }
-            normalized = collapsed
-        }
-        return normalized
-    }
-
-    static func remoteSearchSuggestion(_ query: String) -> OmnibarSuggestion {
-        OmnibarSuggestion(kind: .remote(query: query))
-    }
-
-    private static func displayURLText(for rawURL: String) -> String {
-        guard let components = URLComponents(string: rawURL),
-              var host = components.host else {
-            return rawURL
-        }
-
-        if host.hasPrefix("www.") {
-            host.removeFirst(4)
-        }
-        host = host.lowercased()
-
-        var result = host
-        if let port = components.port {
-            result += ":\(port)"
-        }
-
-        let path = components.percentEncodedPath
-        if !path.isEmpty, path != "/" {
-            result += path
-        } else if path == "/" {
-            result += "/"
-        }
-
-        if let query = components.percentEncodedQuery, !query.isEmpty {
-            result += "?\(query)"
-        }
-
-        if result.isEmpty { return rawURL }
-        return result
-    }
-}
-
-func browserOmnibarShouldReacquireFocusAfterEndEditing(
-    desiredOmnibarFocus: Bool,
-    nextResponderIsOtherTextField: Bool
-) -> Bool {
-    desiredOmnibarFocus && !nextResponderIsOtherTextField
-}
-
-func browserOmnibarShouldSelectAllOnFocusReassertion(
-    selectionIntent: BrowserAddressBarFocusSelectionIntent
-) -> Bool {
-    selectionIntent.shouldSelectAll
-}
-
-/// Whether a completed single click that just moved first responder into the
-/// omnibar should select the field's entire contents (Chrome/Safari/Arc parity),
-/// instead of leaving the caret the field editor placed at the click point.
-///
-/// The first click on an unfocused omnibar showing a URL selects everything so
-/// the user can immediately type a replacement. A subsequent click (the field is
-/// already first responder, so `gainedFocusOnThisClick` is `false`) keeps the
-/// caret placement from https://github.com/manaflow-ai/cmux/issues/5268. A drag
-/// or a Shift-click expresses an explicit range, so select-all defers to it; a
-/// double-click never reaches this path (the field routes multi-clicks straight
-/// to the field editor for word/line selection, and its second click lands after
-/// this click's `mouseUp`, so word selection wins).
-///
-/// - Parameters:
-///   - gainedFocusOnThisClick: `true` when the field had no field editor at
-///     `mouseDown`, i.e. this click is the one that moved focus into the omnibar.
-///   - isShiftClick: `true` when Shift was held, extending an explicit selection.
-///   - didDrag: `true` when the pointer moved far enough to build a drag selection.
-/// - Returns: `true` only for an undragged, unmodified focus-gaining click.
-func browserOmnibarFocusGainingClickShouldSelectAll(
-    gainedFocusOnThisClick: Bool,
-    isShiftClick: Bool,
-    didDrag: Bool
-) -> Bool {
-    gainedFocusOnThisClick && !isShiftClick && !didDrag
-}
-
-final class OmnibarNativeTextField: NSTextField {
-    var panelId: UUID?
-    var onPointerDown: (() -> Void)?
-    var onHandleKeyEvent: ((NSEvent, NSTextView?) -> Bool)?
-    var suppressNextFocusReacquireOnEndEditing = false
-    /// Anchor index for Shift+click selection extension, reset on non-shift clicks.
-    private var shiftClickAnchor: Int?
-    private var mouseSelectionState: MouseSelectionState?
-    private static let dragSelectionThreshold: CGFloat = 3
-
-    private struct MouseSelectionState {
-        let anchor: Int
-        let initialWindowLocation: NSPoint
-        var didDrag: Bool
-        /// `true` when this click moved first responder into the omnibar, gating
-        /// the Chrome-style select-all-on-focus behavior applied at `mouseUp`.
-        let gainedFocus: Bool
-        /// `true` when Shift was held, so an explicit selection extension overrides
-        /// the focus-gaining select-all.
-        let isShift: Bool
-    }
-
-    override init(frame frameRect: NSRect) {
-        super.init(frame: frameRect)
-        isBordered = false
-        isBezeled = false
-        drawsBackground = false
-        focusRingType = .none
-        lineBreakMode = .byTruncatingTail
-        usesSingleLineMode = true
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func resetCursorRects() {
-        super.resetCursorRects()
-        addCursorRect(bounds, cursor: .iBeam)
-    }
-
-    override func mouseDown(with event: NSEvent) {
-        let hadEditor = currentEditor() != nil
-        onPointerDown?()
-
-        if !hadEditor {
-            _ = window?.makeFirstResponder(self)
-        }
-
-        guard let editor = currentEditor() as? NSTextView else {
-            super.mouseDown(with: event)
-            return
-        }
-
-        let isShiftClick = event.modifierFlags.contains(.shift)
-
-        // Keep multi-click word and line selection in the field editor, while avoiding
-        // NSTextField's mouse tracking loop for ordinary clicks.
-        if event.clickCount > 1 {
-            mouseSelectionState = nil
-            editor.mouseDown(with: event)
-            shiftClickAnchor = nil
-            return
-        }
-
-        let clickIndex = insertionIndex(for: event, in: editor)
-        let anchor: Int
-        if isShiftClick {
-            let selected = editor.selectedRange()
-            anchor = shiftClickAnchor ?? selected.location
-            shiftClickAnchor = anchor
-            setSelection(anchor: anchor, extent: clickIndex, in: editor)
-        } else {
-            anchor = clickIndex
-            shiftClickAnchor = nil
-            editor.setSelectedRange(NSRange(location: clickIndex, length: 0))
-        }
-
-        mouseSelectionState = MouseSelectionState(
-            anchor: anchor,
-            initialWindowLocation: event.locationInWindow,
-            didDrag: false,
-            gainedFocus: !hadEditor,
-            isShift: isShiftClick
-        )
-    }
-
-    override func mouseDragged(with event: NSEvent) {
-        guard var state = mouseSelectionState,
-              let editor = currentEditor() as? NSTextView else {
-            super.mouseDragged(with: event)
-            return
-        }
-
-        let dx = event.locationInWindow.x - state.initialWindowLocation.x
-        let dy = event.locationInWindow.y - state.initialWindowLocation.y
-        let distance = (dx * dx + dy * dy).squareRoot()
-        if state.didDrag || distance >= Self.dragSelectionThreshold {
-            state.didDrag = true
-            setSelection(anchor: state.anchor, extent: insertionIndex(for: event, in: editor), in: editor)
-            mouseSelectionState = state
-        }
-    }
-
-    override func mouseUp(with event: NSEvent) {
-        guard let state = mouseSelectionState else {
-            super.mouseUp(with: event)
-            return
-        }
-        mouseSelectionState = nil
-
-        // Chrome/Safari/Arc parity: the click that moves first responder into the
-        // omnibar selects the whole URL so the next keystroke replaces it. A click
-        // while already focused keeps the caret placed in `mouseDown` (issue #5268),
-        // and a drag or Shift-click keeps the explicit range built up during the
-        // gesture. Double-clicks never reach here — `mouseDown` routes multi-clicks
-        // to the field editor for word/line selection and leaves `mouseSelectionState`
-        // nil, and the second click lands after this `mouseUp`, so word selection wins.
-        // The keyboard path (Cmd+L) still selects all via the `selectAllRequestId` flow.
-        guard browserOmnibarFocusGainingClickShouldSelectAll(
-            gainedFocusOnThisClick: state.gainedFocus,
-            isShiftClick: state.isShift,
-            didDrag: state.didDrag
-        ), let editor = currentEditor() as? NSTextView else {
-            return
-        }
-        editor.setSelectedRange(NSRange(location: 0, length: editor.string.utf16.count))
-    }
-
-    private func insertionIndex(for event: NSEvent, in editor: NSTextView) -> Int {
-        let localPoint = editor.convert(event.locationInWindow, from: nil)
-        let index = editor.characterIndexForInsertion(at: localPoint)
-        let textLength = (editor.string as NSString).length
-        guard index != NSNotFound else { return textLength }
-        return min(max(index, 0), textLength)
-    }
-
-    private func setSelection(anchor: Int, extent: Int, in editor: NSTextView) {
-        if extent >= anchor {
-            editor.setSelectedRange(NSRange(location: anchor, length: extent - anchor))
-        } else {
-            editor.setSelectedRange(NSRange(location: extent, length: anchor - extent))
-        }
-    }
-
-    override func keyDown(with event: NSEvent) {
-#if DEBUG
-        let typingTimingStart = CmuxTypingTiming.start()
-        var route = "super"
-        defer {
-            CmuxTypingTiming.logDuration(
-                path: "browser.omnibar.keyDown",
-                startedAt: typingTimingStart,
-                event: event,
-                extra: "route=\(route)"
-            )
-        }
-#endif
-        // Reset shift-click anchor on any keyboard input so that a subsequent
-        // Shift+click uses the post-keyboard selection as its anchor, not a
-        // stale value from a prior mouse interaction.
-        shiftClickAnchor = nil
-        mouseSelectionState = nil
-        if (currentEditor() as? NSTextView)?.hasMarkedText() == true {
-            super.keyDown(with: event)
-            return
-        }
-        if onHandleKeyEvent?(event, currentEditor() as? NSTextView) == true {
-#if DEBUG
-            route = "custom"
-#endif
-            return
-        }
-        super.keyDown(with: event)
-    }
-
-    override func performKeyEquivalent(with event: NSEvent) -> Bool {
-#if DEBUG
-        let typingTimingStart = CmuxTypingTiming.start()
-        var handled = false
-        defer {
-            CmuxTypingTiming.logDuration(
-                path: "browser.omnibar.performKeyEquivalent",
-                startedAt: typingTimingStart,
-                event: event,
-                extra: "handled=\(handled ? 1 : 0)"
-            )
-        }
-#endif
-        shiftClickAnchor = nil
-        mouseSelectionState = nil
-        if (currentEditor() as? NSTextView)?.hasMarkedText() == true {
-            let result = super.performKeyEquivalent(with: event)
-#if DEBUG
-            handled = result
-#endif
-            return result
-        }
-        if onHandleKeyEvent?(event, currentEditor() as? NSTextView) == true {
-#if DEBUG
-            handled = true
-#endif
-            return true
-        }
-        let result = super.performKeyEquivalent(with: event)
-#if DEBUG
-        handled = result
-#endif
-        return result
-    }
-}
-
-struct OmnibarTextFieldRepresentable: NSViewRepresentable {
-    let panelId: UUID
-    let fontSize: CGFloat
-    @Binding var text: String
-    @Binding var isFocused: Bool
-    let selectAllRequestId: UInt64
-    let inlineCompletion: OmnibarInlineCompletion?
-    let placeholder: String
-    let onTap: () -> Void
-    let onSubmit: (OmnibarLiveFieldSnapshot?) -> Void
-    let onEscape: () -> Void
-    let onFieldLostFocus: () -> Void
-    let onMoveSelection: (Int) -> Void
-    let onDeleteSelectedSuggestion: () -> Void
-    let onAcceptInlineCompletion: () -> Void
-    let onDeleteBackwardWithInlineSelection: () -> Void
-    let onClearTypedPrefixWithInlineSelection: () -> Void
-    let onDeleteWordBackwardWithInlineSelection: () -> Void
-    let onSelectionChanged: (NSRange, Bool) -> Void
-    let shouldSuppressWebViewFocus: () -> Bool
-
-    @MainActor
-    final class Coordinator: NSObject, NSTextFieldDelegate {
-        var parent: OmnibarTextFieldRepresentable
-        var isProgrammaticMutation: Bool = false
-        var selectionObserver: NSObjectProtocol?
-        weak var observedEditor: NSTextView?
-        var appliedInlineCompletion: OmnibarInlineCompletion?
-        var lastPublishedSelection: NSRange = NSRange(location: NSNotFound, length: 0)
-        var lastPublishedHasMarkedText: Bool = false
-        /// Guards against infinite focus loops: `true` = focus requested, `false` = blur requested, `nil` = idle.
-        var pendingFocusRequest: Bool?
-        var pendingSelectAllRequestId: UInt64?
-        var appliedSelectAllRequestId: UInt64 = 0
-
-        init(parent: OmnibarTextFieldRepresentable) {
-            self.parent = parent
-        }
-
-#if DEBUG
-        func logFocusEvent(_ event: String, detail: String = "") {
-            let window = parentField?.window
-            let responder = window?.firstResponder
-            let responderType = responder.map { String(describing: type(of: $0)) } ?? "nil"
-            let responderIsField: Int = {
-                guard let field = parentField else { return 0 }
-                if responder === field { return 1 }
-                if let editor = responder as? NSTextView,
-                   (editor.delegate as? NSTextField) === field {
-                    return 1
-                }
-                return 0
-            }()
-            let pendingValue: String = {
-                guard let pendingFocusRequest else { return "nil" }
-                return pendingFocusRequest ? "focus" : "blur"
-            }()
-            var line =
-                "browser.focus.field event=\(event) focused=\(parent.isFocused ? 1 : 0) " +
-                "pending=\(pendingValue) suppressWeb=\(parent.shouldSuppressWebViewFocus() ? 1 : 0) " +
-                "win=\(window?.windowNumber ?? -1) fr=\(responderType) frIsField=\(responderIsField)"
-            if !detail.isEmpty {
-                line += " \(detail)"
-            }
-            cmuxDebugLog(line)
-        }
-#endif
-
-        deinit {
-            if let selectionObserver {
-                NotificationCenter.default.removeObserver(selectionObserver)
-            }
-        }
-
-        private func nextResponderIsOtherTextField(window: NSWindow?) -> Bool {
-            guard let window, let field = parentField else { return false }
-            let responder = window.firstResponder
-
-            if let editor = responder as? NSTextView,
-               let delegateField = editor.delegate as? NSTextField {
-                return delegateField !== field
-            }
-
-            if let textField = responder as? NSTextField {
-                return textField !== field
-            }
-
-            return false
-        }
-
-        private func isPointerDownEvent(_ event: NSEvent) -> Bool {
-            switch event.type {
-            case .leftMouseDown, .rightMouseDown, .otherMouseDown:
-                return true
-            default:
-                return false
-            }
-        }
-
-        private func topHitViewForCurrentPointerEvent(window: NSWindow) -> NSView? {
-            guard let event = NSApp.currentEvent, isPointerDownEvent(event) else {
-                return nil
-            }
-            if event.windowNumber != 0, event.windowNumber != window.windowNumber {
-                return nil
-            }
-            if let eventWindow = event.window, eventWindow !== window {
-                return nil
-            }
-
-            if let contentView = window.contentView,
-               let themeFrame = contentView.superview {
-                let pointInTheme = themeFrame.convert(event.locationInWindow, from: nil)
-                if let hitInTheme = themeFrame.hitTest(pointInTheme) {
-                    return hitInTheme
-                }
-            }
-
-            guard let contentView = window.contentView else {
-                return nil
-            }
-            let pointInContent = contentView.convert(event.locationInWindow, from: nil)
-            return contentView.hitTest(pointInContent)
-        }
-
-        private func pointerDownBlurIntent(window: NSWindow?) -> Bool {
-            guard let window, let field = parentField else { return false }
-            guard let hitView = topHitViewForCurrentPointerEvent(window: window) else {
-                return false
-            }
-
-            if hitView === field || hitView.isDescendant(of: field) {
-                return false
-            }
-            if let interactionView = hitView as? BrowserOmnibarInteractionView,
-               interactionView.panelId == field.panelId {
-                return false
-            }
-            if let textView = hitView as? NSTextView,
-               let delegateField = textView.delegate as? NSTextField,
-               delegateField === field {
-                return false
-            }
-            return true
-        }
-
-        private func shouldReacquireFocusAfterEndEditing(window: NSWindow?) -> Bool {
-            if parentField?.suppressNextFocusReacquireOnEndEditing == true {
-                return false
-            }
-            if pointerDownBlurIntent(window: window) {
-                return false
-            }
-            return browserOmnibarShouldReacquireFocusAfterEndEditing(
-                desiredOmnibarFocus: parent.isFocused,
-                nextResponderIsOtherTextField: nextResponderIsOtherTextField(window: window)
-            )
-        }
-
-        func controlTextDidBeginEditing(_ obj: Notification) {
-#if DEBUG
-            logFocusEvent("controlTextDidBeginEditing")
-#endif
-            if !parent.isFocused {
-                DispatchQueue.main.async {
-#if DEBUG
-                    self.logFocusEvent("controlTextDidBeginEditing.asyncSetFocused", detail: "old=0 new=1")
-#endif
-                    self.parent.isFocused = true
-                }
-            }
-            attachSelectionObserverIfNeeded()
-            if let field = obj.object as? OmnibarNativeTextField {
-                field.suppressNextFocusReacquireOnEndEditing = false
-                applyPendingSelectAllIfPossible(field: field)
-            }
-            publishSelectionState()
-        }
-
-        func controlTextDidEndEditing(_ obj: Notification) {
-            let shouldReacquire = shouldReacquireFocusAfterEndEditing(window: parentField?.window)
-#if DEBUG
-            let nextOther = nextResponderIsOtherTextField(window: parentField?.window)
-            let pointerBlur = pointerDownBlurIntent(window: parentField?.window)
-            logFocusEvent(
-                "controlTextDidEndEditing",
-                detail: "nextOther=\(nextOther ? 1 : 0) pointerBlur=\(pointerBlur ? 1 : 0) shouldReacquire=\(shouldReacquire ? 1 : 0)"
-            )
-#endif
-            if parent.isFocused {
-                if shouldReacquire {
-#if DEBUG
-                    logFocusEvent("controlTextDidEndEditing.reacquire.begin")
-#endif
-                    guard pendingFocusRequest != true else { return }
-                    pendingFocusRequest = true
-                    DispatchQueue.main.async { [weak self] in
-                        guard let self else { return }
-                        self.pendingFocusRequest = nil
-#if DEBUG
-                        self.logFocusEvent("controlTextDidEndEditing.reacquire.tick")
-#endif
-                        guard self.parent.isFocused else { return }
-                        guard let field = self.parentField, let window = field.window else { return }
-                        guard self.shouldReacquireFocusAfterEndEditing(window: window) else {
-#if DEBUG
-                            self.logFocusEvent("controlTextDidEndEditing.reacquire.cancel")
-#endif
-                            self.parent.onFieldLostFocus()
-                            return
-                        }
-                        // Check both the field itself AND its field editor (which becomes
-                        // the actual first responder when the text field is being edited).
-                        let fr = window.firstResponder
-                        let isAlreadyFocused = fr === field ||
-                            field.currentEditor() != nil ||
-                            ((fr as? NSTextView)?.delegate as? NSTextField) === field
-                        if !isAlreadyFocused {
-#if DEBUG
-                            self.logFocusEvent("controlTextDidEndEditing.reacquire.apply")
-#endif
-                            window.makeFirstResponder(field)
-                        } else {
-#if DEBUG
-                            self.logFocusEvent("controlTextDidEndEditing.reacquire.skip", detail: "reason=already_focused")
-#endif
-                        }
-                    }
-                    return
-                }
-#if DEBUG
-                logFocusEvent("controlTextDidEndEditing.blur")
-#endif
-                parent.onFieldLostFocus()
-            }
-            parentField?.suppressNextFocusReacquireOnEndEditing = false
-            detachSelectionObserver()
-        }
-
-        func controlTextDidChange(_ obj: Notification) {
-#if DEBUG
-            let typingTimingStart = CmuxTypingTiming.start()
-            defer {
-                CmuxTypingTiming.logDuration(
-                    path: "browser.omnibar.controlTextDidChange",
-                    startedAt: typingTimingStart,
-                    event: NSApp.currentEvent,
-                    extra: "programmatic=\(isProgrammaticMutation ? 1 : 0)"
-                )
-            }
-#endif
-            guard !isProgrammaticMutation else { return }
-            guard let field = obj.object as? NSTextField else { return }
-            let editor = field.currentEditor() as? NSTextView
-            publishSelectionState()
-            parent.text = omnibarPublishedBufferTextForFieldChange(
-                fieldValue: field.stringValue,
-                inlineCompletion: parent.inlineCompletion,
-                selectionRange: editor?.selectedRange(),
-                hasMarkedText: editor?.hasMarkedText() ?? false
-            )
-        }
-
-        func control(_ control: NSControl, textView: NSTextView, doCommandBy commandSelector: Selector) -> Bool {
-#if DEBUG
-            let typingTimingStart = CmuxTypingTiming.start()
-            var handled = false
-            defer {
-                CmuxTypingTiming.logDuration(
-                    path: "browser.omnibar.doCommandBy",
-                    startedAt: typingTimingStart,
-                    event: NSApp.currentEvent,
-                    extra: "handled=\(handled ? 1 : 0) selector=\(NSStringFromSelector(commandSelector))"
-                )
-            }
-#endif
-            guard !textView.hasMarkedText() else { return false }
-            switch commandSelector {
-            case #selector(NSResponder.moveDown(_:)):
-                parent.onMoveSelection(+1)
-#if DEBUG
-                handled = true
-#endif
-                return true
-            case #selector(NSResponder.moveUp(_:)):
-                parent.onMoveSelection(-1)
-#if DEBUG
-                handled = true
-#endif
-                return true
-            case #selector(NSResponder.insertNewline(_:)):
-                let currentFlags = NSApp.currentEvent?.modifierFlags ?? []
-                guard browserOmnibarShouldSubmitOnReturn(flags: currentFlags) else { return false }
-                parent.onSubmit(liveFieldSnapshot(preferredEditor: textView))
-#if DEBUG
-                handled = true
-#endif
-                return true
-            case #selector(NSResponder.cancelOperation(_:)):
-                parent.onEscape()
-#if DEBUG
-                handled = true
-#endif
-                return true
-            case #selector(NSResponder.moveRight(_:)), #selector(NSResponder.moveToEndOfLine(_:)):
-                if parent.inlineCompletion != nil {
-                    parent.onAcceptInlineCompletion()
-#if DEBUG
-                    handled = true
-#endif
-                    return true
-                }
-                return false
-            case #selector(NSResponder.insertTab(_:)):
-                if parent.inlineCompletion != nil {
-                    parent.onAcceptInlineCompletion()
-#if DEBUG
-                    handled = true
-#endif
-                    return true
-                }
-                return false
-            case #selector(NSResponder.deleteBackward(_:)):
-                if suffixSelectionMatchesInline(textView, inline: parent.inlineCompletion) {
-                    parent.onDeleteBackwardWithInlineSelection()
-#if DEBUG
-                    handled = true
-#endif
-                    return true
-                }
-                return false
-            case #selector(NSResponder.deleteToBeginningOfLine(_:)),
-                 #selector(NSResponder.deleteToBeginningOfParagraph(_:)):
-                if inlineCompletionSelectionIsActive(textView, inline: parent.inlineCompletion) {
-                    parent.onClearTypedPrefixWithInlineSelection()
-#if DEBUG
-                    handled = true
-#endif
-                    return true
-                }
-                return false
-            case #selector(NSResponder.deleteWordBackward(_:)):
-                if inlineCompletionSelectionIsActive(textView, inline: parent.inlineCompletion) {
-                    parent.onDeleteWordBackwardWithInlineSelection()
-#if DEBUG
-                    handled = true
-#endif
-                    return true
-                }
-                return false
-            default:
-                return false
-            }
-        }
-
-        func attachSelectionObserverIfNeeded() {
-            guard selectionObserver == nil else { return }
-            guard let field = parentField else { return }
-            guard let editor = field.currentEditor() as? NSTextView else { return }
-            observedEditor = editor
-            selectionObserver = NotificationCenter.default.addObserver(
-                forName: NSTextView.didChangeSelectionNotification,
-                object: editor,
-                queue: .main
-            ) { [weak self] _ in
-                MainActor.assumeIsolated {
-                    self?.publishSelectionState()
-                }
-            }
-        }
-
-        func detachSelectionObserver() {
-            if let selectionObserver {
-                NotificationCenter.default.removeObserver(selectionObserver)
-                self.selectionObserver = nil
-            }
-            observedEditor = nil
-        }
-
-        weak var parentField: OmnibarNativeTextField?
-
-        func queueSelectAllRequest(_ requestId: UInt64) {
-            guard requestId != 0, appliedSelectAllRequestId != requestId else { return }
-            pendingSelectAllRequestId = requestId
-        }
-
-        @discardableResult
-        func applyPendingSelectAllIfPossible(
-            field: OmnibarNativeTextField
-        ) -> Bool {
-            guard let requestId = pendingSelectAllRequestId,
-                  requestId != 0,
-                  appliedSelectAllRequestId != requestId else {
-                return false
-            }
-
-            guard let editor = field.currentEditor() as? NSTextView,
-                  !editor.hasMarkedText() else {
-                return false
-            }
-            let length = editor.string.utf16.count
-            isProgrammaticMutation = true
-            editor.setSelectedRange(NSRange(location: 0, length: length))
-            isProgrammaticMutation = false
-            appliedSelectAllRequestId = requestId
-            pendingSelectAllRequestId = nil
-            publishSelectionState()
-            return true
-        }
-
-        func publishSelectionState() {
-            guard let field = parentField else { return }
-            if let editor = field.currentEditor() as? NSTextView {
-                let range = editor.selectedRange()
-                let hasMarkedText = editor.hasMarkedText()
-                guard !NSEqualRanges(range, lastPublishedSelection) || hasMarkedText != lastPublishedHasMarkedText else {
-                    return
-                }
-                lastPublishedSelection = range
-                lastPublishedHasMarkedText = hasMarkedText
-                parent.onSelectionChanged(range, hasMarkedText)
-            } else {
-                let location = field.stringValue.utf16.count
-                let range = NSRange(location: location, length: 0)
-                guard !NSEqualRanges(range, lastPublishedSelection) || lastPublishedHasMarkedText else { return }
-                lastPublishedSelection = range
-                lastPublishedHasMarkedText = false
-                parent.onSelectionChanged(range, false)
-            }
-        }
-
-        /// Captures the field-editor text synchronously at submit time. Both
-        /// Return interception paths (`doCommandBy` and `handleKeyEvent`) go
-        /// through here so the submit decision always starts from what the
-        /// field actually shows, not the possibly lagging published state.
-        private func liveFieldSnapshot(preferredEditor: NSTextView?) -> OmnibarLiveFieldSnapshot? {
-            let editor = preferredEditor ?? (parentField?.currentEditor() as? NSTextView)
-            if let editor {
-                return OmnibarLiveFieldSnapshot(
-                    text: editor.string,
-                    selectionRange: editor.selectedRange(),
-                    hasMarkedText: editor.hasMarkedText()
-                )
-            }
-            guard let field = parentField else { return nil }
-            return OmnibarLiveFieldSnapshot(
-                text: field.stringValue,
-                selectionRange: nil,
-                hasMarkedText: false
-            )
-        }
-
-        private func inlineCompletionSelectionIsActive(_ editor: NSTextView?, inline: OmnibarInlineCompletion?) -> Bool {
-            suffixSelectionMatchesInline(editor, inline: inline) || selectionIsTypedPrefixBoundary(editor, inline: inline)
-        }
-
-        private func suffixSelectionMatchesInline(_ editor: NSTextView?, inline: OmnibarInlineCompletion?) -> Bool {
-            guard let editor, let inline else { return false }
-            let selected = editor.selectedRange()
-            return NSEqualRanges(selected, inline.suffixRange)
-        }
-
-        private func selectionIsTypedPrefixBoundary(_ editor: NSTextView?, inline: OmnibarInlineCompletion?) -> Bool {
-            guard let editor, let inline else { return false }
-            let selected = editor.selectedRange()
-            let typedCount = inline.typedText.utf16.count
-            return selected.location == typedCount && selected.length == 0
-        }
-
-        func handleKeyEvent(_ event: NSEvent, editor: NSTextView?) -> Bool {
-#if DEBUG
-            let typingTimingStart = CmuxTypingTiming.start()
-            var handled = false
-            defer {
-                CmuxTypingTiming.logDuration(
-                    path: "browser.omnibar.handleKeyEvent",
-                    startedAt: typingTimingStart,
-                    event: event,
-                    extra: "handled=\(handled ? 1 : 0)"
-                )
-            }
-#endif
-            // #6250: AppKit invokes `performKeyEquivalent` across the entire
-            // window view hierarchy, so this coordinator runs even while web
-            // content (the WKWebView) — not the omnibar — owns first responder.
-            // In that state the omnibar field has no field editor, so `editor`
-            // is nil. Treating Return/Escape/arrows (and Ctrl+N/P, Shift+Delete)
-            // as omnibar input there makes an *unfocused* omnibar submit and
-            // hard-navigate the pane on a physical Enter that belongs to the
-            // page — a spurious reload that aborts in-flight `fetch`/XHR in SPAs.
-            // Only act on these keys while the field is actually being edited.
-            // This mirrors the `currentEditor()`-gated `insertNewline:` path in
-            // `control(_:textView:doCommandBy:)`, which only runs for the live
-            // field editor.
-            guard editor != nil else { return false }
-            guard editor?.hasMarkedText() != true else { return false }
-            let keyCode = event.keyCode
-            let modifiers = event.modifierFlags.intersection([.command, .control, .shift, .option, .function])
-            // When a non-Latin input source is active (Korean, Chinese, Japanese),
-            // charactersIgnoringModifiers returns non-ASCII characters. Normalize
-            // via KeyboardLayout so Ctrl+N/P navigation works across input sources.
-            let lowered = KeyboardLayout.normalizedCharacters(for: event)
-
-            // Ctrl+N and Ctrl+P should repeat while held.
-            if let delta = browserOmnibarSelectionDeltaForControlNavigation(
-                hasFocusedAddressBar: true,
-                flags: event.modifierFlags,
-                chars: lowered
-            ) {
-                parent.onMoveSelection(delta)
-#if DEBUG
-                handled = true
-#endif
-                return true
-            }
-
-            // Shift+Delete removes the selected history suggestion when possible.
-            if modifiers.contains(.shift), (keyCode == 51 || keyCode == 117) {
-                parent.onDeleteSelectedSuggestion()
-#if DEBUG
-                handled = true
-#endif
-                return true
-            }
-
-            switch keyCode {
-            case 36, 76: // Return / keypad Enter
-                guard browserOmnibarShouldSubmitOnReturn(flags: event.modifierFlags) else { return false }
-                parent.onSubmit(liveFieldSnapshot(preferredEditor: editor))
-#if DEBUG
-                handled = true
-#endif
-                return true
-            case 53: // Escape
-                parent.onEscape()
-#if DEBUG
-                handled = true
-#endif
-                return true
-            case 125: // Down
-                parent.onMoveSelection(+1)
-#if DEBUG
-                handled = true
-#endif
-                return true
-            case 126: // Up
-                parent.onMoveSelection(-1)
-#if DEBUG
-                handled = true
-#endif
-                return true
-            case 124, 119: // Right arrow / End
-                if parent.inlineCompletion != nil {
-                    parent.onAcceptInlineCompletion()
-#if DEBUG
-                    handled = true
-#endif
-                    return true
-                }
-            case 48: // Tab
-                if parent.inlineCompletion != nil {
-                    parent.onAcceptInlineCompletion()
-#if DEBUG
-                    handled = true
-#endif
-                    return true
-                }
-            case 51: // Backspace
-                if modifiers.contains(.command) || modifiers.contains(.option) {
-                    return false
-                }
-                if let inline = parent.inlineCompletion,
-                   inlineCompletionSelectionIsActive(editor, inline: inline) {
-                    parent.onDeleteBackwardWithInlineSelection()
-#if DEBUG
-                    handled = true
-#endif
-                    return true
-                }
-            default:
-                break
-            }
-
-            return false
-        }
-    }
-
-    func makeCoordinator() -> Coordinator {
-        Coordinator(parent: self)
-    }
-
-    func makeNSView(context: Context) -> OmnibarNativeTextField {
-        let field = OmnibarNativeTextField(frame: .zero)
-        field.panelId = panelId
-        BrowserOmnibarNativeFieldRegistry.shared.register(field, panelId: panelId)
-        field.identifier = browserOmnibarTextFieldIdentifier
-        field.font = .systemFont(ofSize: fontSize)
-        field.placeholderString = placeholder
-        field.delegate = context.coordinator
-        field.target = nil
-        field.action = nil
-        field.isEditable = true
-        field.isSelectable = true
-        field.isEnabled = true
-        field.stringValue = text
-        field.onPointerDown = {
-            onTap()
-        }
-        field.onHandleKeyEvent = { [weak coordinator = context.coordinator] event, editor in
-            coordinator?.handleKeyEvent(event, editor: editor) ?? false
-        }
-        context.coordinator.parentField = field
-        BrowserOmnibarNativeFieldRegistry.shared.register(field, panelId: panelId)
-        return field
-    }
-
-    func updateNSView(_ nsView: OmnibarNativeTextField, context: Context) {
-        context.coordinator.parent = self
-        context.coordinator.parentField = nsView
-        if let previousPanelId = nsView.panelId, previousPanelId != panelId {
-            BrowserOmnibarNativeFieldRegistry.shared.unregister(nsView, panelId: previousPanelId)
-        }
-        nsView.panelId = panelId
-        BrowserOmnibarNativeFieldRegistry.shared.register(nsView, panelId: panelId)
-        nsView.placeholderString = placeholder
-        if nsView.font?.pointSize != fontSize {
-            nsView.font = .systemFont(ofSize: fontSize)
-        }
-        context.coordinator.queueSelectAllRequest(selectAllRequestId)
-
-        let activeInlineCompletion = omnibarInlineCompletionIfBufferMatchesTypedPrefix(
-            bufferText: text,
-            inlineCompletion: inlineCompletion
-        )
-        let desiredDisplayText = activeInlineCompletion?.displayText ?? text
-        if let editor = nsView.currentEditor() as? NSTextView {
-            if !editor.hasMarkedText(), editor.string != desiredDisplayText {
-                context.coordinator.isProgrammaticMutation = true
-                editor.string = desiredDisplayText
-                nsView.stringValue = desiredDisplayText
-                context.coordinator.isProgrammaticMutation = false
-            }
-        } else if nsView.stringValue != desiredDisplayText {
-            nsView.stringValue = desiredDisplayText
-        }
-
-        if let window = nsView.window {
-            let firstResponder = window.firstResponder
-            let isFirstResponder =
-                firstResponder === nsView ||
-                nsView.currentEditor() != nil ||
-                ((firstResponder as? NSTextView)?.delegate as? NSTextField) === nsView
-            if isFocused, !isFirstResponder, context.coordinator.pendingFocusRequest != true {
-#if DEBUG
-                context.coordinator.logFocusEvent(
-                    "updateNSView.requestFocus.begin",
-                    detail: "isFocused=1 isFirstResponder=0"
-                )
-#endif
-                // Defer to avoid triggering input method XPC during layout pass,
-                // which can crash via re-entrant view hierarchy modification.
-                context.coordinator.pendingFocusRequest = true
-                DispatchQueue.main.async { [weak nsView, weak coordinator = context.coordinator] in
-                    coordinator?.pendingFocusRequest = nil
-                    guard let nsView, let window = nsView.window else { return }
-#if DEBUG
-                    if coordinator?.parent.isFocused != true {
-                        coordinator?.logFocusEvent("updateNSView.requestFocus.cancel", detail: "reason=stale_state")
-                        return
-                    }
-#endif
-                    guard coordinator?.parent.isFocused == true else { return }
-#if DEBUG
-                    coordinator?.logFocusEvent("updateNSView.requestFocus.tick")
-#endif
-                    let fr = window.firstResponder
-                    let alreadyFocused = fr === nsView ||
-                        nsView.currentEditor() != nil ||
-                        ((fr as? NSTextView)?.delegate as? NSTextField) === nsView
-                    if alreadyFocused {
-                        coordinator?.applyPendingSelectAllIfPossible(field: nsView)
-                        return
-                    }
-#if DEBUG
-                    coordinator?.logFocusEvent("updateNSView.requestFocus.apply")
-#endif
-                    window.makeFirstResponder(nsView)
-                    coordinator?.applyPendingSelectAllIfPossible(field: nsView)
-                }
-            } else if !isFocused, isFirstResponder, context.coordinator.pendingFocusRequest != false {
-#if DEBUG
-                context.coordinator.logFocusEvent(
-                    "updateNSView.requestBlur.begin",
-                    detail: "isFocused=0 isFirstResponder=1"
-                )
-#endif
-                context.coordinator.pendingFocusRequest = false
-                DispatchQueue.main.async { [weak nsView, weak coordinator = context.coordinator] in
-                    coordinator?.pendingFocusRequest = nil
-                    guard let nsView, let window = nsView.window else { return }
-#if DEBUG
-                    if coordinator?.parent.isFocused == true {
-                        coordinator?.logFocusEvent("updateNSView.requestBlur.cancel", detail: "reason=stale_state")
-                        return
-                    }
-#endif
-                    guard coordinator?.parent.isFocused == false else { return }
-#if DEBUG
-                    coordinator?.logFocusEvent("updateNSView.requestBlur.tick")
-#endif
-                    let fr = window.firstResponder
-                    let stillFirst = fr === nsView ||
-                        ((fr as? NSTextView)?.delegate as? NSTextField) === nsView
-                    guard stillFirst else { return }
-#if DEBUG
-                    coordinator?.logFocusEvent("updateNSView.requestBlur.apply")
-#endif
-                    window.makeFirstResponder(nil)
-                }
-            }
-        }
-        context.coordinator.applyPendingSelectAllIfPossible(field: nsView)
-
-        if let editor = nsView.currentEditor() as? NSTextView, !editor.hasMarkedText() {
-            if let activeInlineCompletion {
-                let currentSelection = editor.selectedRange()
-                let desiredSelection = omnibarDesiredSelectionRangeForInlineCompletion(
-                    currentSelection: currentSelection,
-                    inlineCompletion: activeInlineCompletion
-                )
-                if context.coordinator.appliedInlineCompletion != activeInlineCompletion ||
-                    !NSEqualRanges(currentSelection, desiredSelection) {
-                    context.coordinator.isProgrammaticMutation = true
-                    editor.setSelectedRange(desiredSelection)
-                    context.coordinator.isProgrammaticMutation = false
-                }
-            } else if context.coordinator.appliedInlineCompletion != nil {
-                let end = text.utf16.count
-                let current = editor.selectedRange()
-                if current.length != 0 || current.location != end {
-                    context.coordinator.isProgrammaticMutation = true
-                    editor.setSelectedRange(NSRange(location: end, length: 0))
-                    context.coordinator.isProgrammaticMutation = false
-                }
-            }
-        }
-        context.coordinator.appliedInlineCompletion = activeInlineCompletion
-        context.coordinator.attachSelectionObserverIfNeeded()
-        context.coordinator.publishSelectionState()
-    }
-
-    static func dismantleNSView(_ nsView: OmnibarNativeTextField, coordinator: Coordinator) {
-        if let panelId = nsView.panelId {
-            BrowserOmnibarNativeFieldRegistry.shared.unregister(nsView, panelId: panelId)
-        }
-        nsView.onPointerDown = nil
-        nsView.onHandleKeyEvent = nil
-        nsView.delegate = nil
-        coordinator.detachSelectionObserver()
-        coordinator.parentField = nil
-    }
-}
-
-@MainActor
-func browserOmnibarPanelId(for responder: NSResponder?) -> UUID? {
-    browserOmnibarField(for: responder)?.panelId
-}
-
-@MainActor
-func browserOmnibarField(panelId: UUID?, in window: NSWindow?) -> OmnibarNativeTextField? {
-    if let registeredField = BrowserOmnibarNativeFieldRegistry.shared.field(for: panelId, in: window) {
-        return registeredField
-    }
-    guard let panelId, let root = window?.contentView?.superview ?? window?.contentView else {
-        return nil
-    }
-
-    // Fallback for SwiftUI/AppKit reconnect windows where the live native field
-    // has been attached but registration has not yet observed it.
-    var stack: [NSView] = [root]
-    while let view = stack.popLast() {
-        if let field = view as? OmnibarNativeTextField, field.panelId == panelId {
-            return field
-        }
-        stack.append(contentsOf: view.subviews)
-    }
-    return nil
-}
-
-@discardableResult
-@MainActor
-func browserPrepareOmnibarForProgrammaticBlur(panelId: UUID, responder: NSResponder?) -> Bool {
-    guard let field = browserOmnibarField(for: responder),
-          field.panelId == panelId else {
-        return false
-    }
-    field.suppressNextFocusReacquireOnEndEditing = true
-    return true
-}
-
-@MainActor
-private func browserOmnibarField(for responder: NSResponder?) -> OmnibarNativeTextField? {
-    guard let responder else { return nil }
-
-    if let field = responder as? OmnibarNativeTextField {
-        return field
-    }
-
-    if let editor = responder as? NSTextView, editor.isFieldEditor {
-        if let field = BrowserOmnibarNativeFieldRegistry.shared.fieldOwningEditor(editor, in: editor.window) {
-            return field
-        }
-
-        if let field = cmuxFieldEditorOwnerView(editor) as? OmnibarNativeTextField,
-           field.currentEditor() === editor {
-            return field
-        }
-
-    }
-
-    return nil
-}
-
-struct OmnibarSuggestionsView: View {
-    let engineName: String
-    let items: [OmnibarSuggestion]
-    let selectedIndex: Int
-    let isLoadingRemoteSuggestions: Bool
-    let searchSuggestionsEnabled: Bool
-    let onCommit: (OmnibarSuggestion) -> Void
-    let onHighlight: (Int) -> Void
-    @Environment(\.colorScheme) private var colorScheme
-
-    // Keep radii below half of the smallest rendered heights so this keeps a
-    // squircle silhouette instead of auto-clamping into a capsule.
-    private static let popupCornerRadius: CGFloat = 12
-    private static let rowHighlightCornerRadius: CGFloat = 9
-    private static let singleLineRowHeight: CGFloat = 24
-    private static let rowSpacing: CGFloat = 1
-    private static let topInset: CGFloat = 3
-    private static let bottomInset: CGFloat = 3
-    private static let maxPopupHeight: CGFloat = 560
-
-    private var popupCornerRadius: CGFloat { Self.popupCornerRadius }
-    private var rowHighlightCornerRadius: CGFloat { Self.rowHighlightCornerRadius }
-    private var singleLineRowHeight: CGFloat { Self.singleLineRowHeight }
-    private var rowSpacing: CGFloat { Self.rowSpacing }
-    private var topInset: CGFloat { Self.topInset }
-    private var bottomInset: CGFloat { Self.bottomInset }
-    private var horizontalInset: CGFloat { topInset }
-    private var maxPopupHeight: CGFloat { Self.maxPopupHeight }
-
-    private var totalRowCount: Int {
-        max(1, items.count)
-    }
-
-    private func rowHeight(for item: OmnibarSuggestion) -> CGFloat {
-        return singleLineRowHeight
-    }
-
-    private var contentHeight: CGFloat {
-        let rowsHeight = items.isEmpty ? singleLineRowHeight : items.reduce(CGFloat(0)) { partial, item in
-            partial + rowHeight(for: item)
-        }
-        let gaps = CGFloat(max(0, totalRowCount - 1))
-        return rowsHeight + (gaps * rowSpacing) + topInset + bottomInset
-    }
-
-    private var minimumPopupHeight: CGFloat {
-        singleLineRowHeight + topInset + bottomInset
-    }
-
-    private func snapToDevicePixels(_ value: CGFloat) -> CGFloat {
-        let scale = NSScreen.main?.backingScaleFactor ?? 2
-        return (value * scale).rounded(.toNearestOrAwayFromZero) / scale
-    }
-
-    private var popupHeight: CGFloat {
-        Self.popupHeight(for: items)
-    }
-
-    static func popupHeight(for items: [OmnibarSuggestion]) -> CGFloat {
-        let totalRowCount = max(1, items.count)
-        let rowsHeight = items.isEmpty ? singleLineRowHeight : CGFloat(items.count) * singleLineRowHeight
-        let gaps = CGFloat(max(0, totalRowCount - 1))
-        let contentHeight = rowsHeight + (gaps * rowSpacing) + topInset + bottomInset
-        let minimumPopupHeight = singleLineRowHeight + topInset + bottomInset
-        let scale = NSScreen.main?.backingScaleFactor ?? 2
-        let height = min(max(contentHeight, minimumPopupHeight), maxPopupHeight)
-        return (height * scale).rounded(.toNearestOrAwayFromZero) / scale
-    }
-
-    private var isPointerDrivenSelectionEvent: Bool {
-        guard let event = NSApp.currentEvent else { return false }
-        switch event.type {
-        case .mouseMoved, .leftMouseDown, .leftMouseDragged, .leftMouseUp,
-             .rightMouseDown, .rightMouseDragged, .rightMouseUp,
-             .otherMouseDown, .otherMouseDragged, .otherMouseUp, .scrollWheel:
-            return true
-        default:
-            return false
-        }
-    }
-
-    private var shouldScroll: Bool {
-        contentHeight > maxPopupHeight
-    }
-
-    private var listTextColor: Color {
-        switch colorScheme {
-        case .light:
-            return Color(nsColor: .labelColor)
-        case .dark:
-            return Color.white.opacity(0.9)
-        @unknown default:
-            return Color(nsColor: .labelColor)
-        }
-    }
-
-    private var badgeTextColor: Color {
-        switch colorScheme {
-        case .light:
-            return Color(nsColor: .secondaryLabelColor)
-        case .dark:
-            return Color.white.opacity(0.72)
-        @unknown default:
-            return Color(nsColor: .secondaryLabelColor)
-        }
-    }
-
-    private var badgeBackgroundColor: Color {
-        switch colorScheme {
-        case .light:
-            return Color.black.opacity(0.06)
-        case .dark:
-            return Color.white.opacity(0.08)
-        @unknown default:
-            return Color.black.opacity(0.06)
-        }
-    }
-
-    private var rowHighlightColor: Color {
-        switch colorScheme {
-        case .light:
-            return Color.black.opacity(0.07)
-        case .dark:
-            return Color.white.opacity(0.12)
-        @unknown default:
-            return Color.black.opacity(0.07)
-        }
-    }
-
-    private var popupOverlayGradientColors: [Color] {
-        switch colorScheme {
-        case .light:
-            return [
-                Color.white.opacity(0.55),
-                Color.white.opacity(0.2),
-            ]
-        case .dark:
-            return [
-                Color.black.opacity(0.26),
-                Color.black.opacity(0.14),
-            ]
-        @unknown default:
-            return [
-                Color.white.opacity(0.55),
-                Color.white.opacity(0.2),
-            ]
-        }
-    }
-
-    private var popupBorderGradientColors: [Color] {
-        switch colorScheme {
-        case .light:
-            return [
-                Color.white.opacity(0.65),
-                Color.black.opacity(0.12),
-            ]
-        case .dark:
-            return [
-                Color.white.opacity(0.22),
-                Color.white.opacity(0.06),
-            ]
-        @unknown default:
-            return [
-                Color.white.opacity(0.65),
-                Color.black.opacity(0.12),
-            ]
-        }
-    }
-
-    private var popupShadowColor: Color {
-        switch colorScheme {
-        case .light:
-            return Color.black.opacity(0.18)
-        case .dark:
-            return Color.black.opacity(0.45)
-        @unknown default:
-            return Color.black.opacity(0.18)
-        }
-    }
-
-    @ViewBuilder
-    private var rowsView: some View {
-        VStack(spacing: rowSpacing) {
-            ForEach(Array(items.enumerated()), id: \.element.id) { idx, item in
-            Button {
-                #if DEBUG
-                let suggestionKind: String = {
-                    switch item.kind {
-                    case .search:
-                        return "search"
-                    case .navigate:
-                        return "navigate"
-                    case .history:
-                        return "history"
-                    case .switchToTab:
-                        return "switchToTab"
-                    case .remote:
-                        return "remote"
-                    }
-                }()
-                cmuxDebugLog("browser.suggestionClick index=\(idx) kind=\(suggestionKind) textBytes=\(item.listText.utf8.count)")
-                #endif
-                onCommit(item)
-            } label: {
-                HStack(spacing: 6) {
-                        Text(item.listText)
-                            .cmuxFont(size: 11)
-                            .foregroundStyle(listTextColor)
-                            .lineLimit(1)
-                            .truncationMode(.tail)
-                        if let badge = item.trailingBadgeText {
-                            Text(badge)
-                                .cmuxFont(size: 9.5, weight: .medium)
-                                .foregroundStyle(badgeTextColor)
-                                .padding(.horizontal, 6)
-                                .padding(.vertical, 2)
-                                .background(
-                                    RoundedRectangle(cornerRadius: 7, style: .continuous)
-                                        .fill(badgeBackgroundColor)
-                                )
-                        }
-                        Spacer(minLength: 0)
-                    }
-                    .padding(.horizontal, 8)
-                    .frame(
-                        maxWidth: .infinity,
-                        minHeight: rowHeight(for: item),
-                        maxHeight: rowHeight(for: item),
-                        alignment: .leading
-                    )
-                    .background(
-                        RoundedRectangle(cornerRadius: rowHighlightCornerRadius, style: .continuous)
-                            .fill(
-                                idx == selectedIndex
-                                    ? rowHighlightColor
-                                    : Color.clear
-                            )
-                    )
-                }
-                .buttonStyle(.plain)
-                .accessibilityIdentifier("BrowserOmnibarSuggestions.Row.\(idx)")
-                .accessibilityValue(
-                    idx == selectedIndex
-                        ? "selected \(item.listText)"
-                        : item.listText
-                )
-                .onHover { hovering in
-                    if hovering, idx != selectedIndex, isPointerDrivenSelectionEvent {
-                        onHighlight(idx)
-                    }
-                }
-                .animation(.none, value: selectedIndex)
-            }
-
-        }
-        .padding(.horizontal, horizontalInset)
-        .padding(.top, topInset)
-        .padding(.bottom, bottomInset)
-        .frame(maxWidth: .infinity, alignment: .topLeading)
-    }
-
-    var body: some View {
-        Group {
-            if shouldScroll {
-                ScrollView {
-                    rowsView
-                }
-            } else {
-                rowsView
-            }
-        }
-        .frame(height: popupHeight, alignment: .top)
-        .overlay(alignment: .topTrailing) {
-            if searchSuggestionsEnabled, isLoadingRemoteSuggestions {
-                ProgressView()
-                    .controlSize(.small)
-                    .padding(.top, 7)
-                    .padding(.trailing, 14)
-                    .opacity(0.75)
-                    .allowsHitTesting(false)
-            }
-        }
-        .background(
-            RoundedRectangle(cornerRadius: popupCornerRadius, style: .continuous)
-                .fill(.ultraThinMaterial)
-                .overlay(
-                    RoundedRectangle(cornerRadius: popupCornerRadius, style: .continuous)
-                        .fill(
-                            LinearGradient(
-                                colors: popupOverlayGradientColors,
-                                startPoint: .top,
-                                endPoint: .bottom
-                            )
-                        )
-                )
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: popupCornerRadius, style: .continuous)
-                .stroke(
-                    LinearGradient(
-                        colors: popupBorderGradientColors,
-                        startPoint: .top,
-                        endPoint: .bottom
-                    ),
-                    lineWidth: 1
-                )
-        )
-        .clipShape(RoundedRectangle(cornerRadius: popupCornerRadius, style: .continuous))
-        .shadow(color: popupShadowColor, radius: 20, y: 10)
-        .contentShape(RoundedRectangle(cornerRadius: popupCornerRadius, style: .continuous))
-        .accessibilityElement(children: .contain)
-        .accessibilityRespondsToUserInteraction(true)
-        .accessibilityIdentifier("BrowserOmnibarSuggestions")
-        .accessibilityLabel(String(localized: "browser.addressBarSuggestions", defaultValue: "Address bar suggestions"))
-    }
-}
-
-/// NSViewRepresentable wrapper for WKWebView
 struct WebViewRepresentable: NSViewRepresentable {
     let panel: BrowserPanel
     let paneId: PaneID
