+++ Sources/Panels/CEFBrowserPanelView.swift
@@ -1,5 +1,4 @@
 import AppKit
-import CEFKit
 import SwiftUI
 
 struct CEFBrowserPanelView: View {
@@ -7,148 +6,90 @@ struct CEFBrowserPanelView: View {
     let isFocused: Bool
     let onRequestPanelFocus: () -> Void
 
-    @FocusState private var isAddressFieldFocused: Bool
+    @Environment(\.colorScheme) private var colorScheme
+    @State private var suggestions: BrowserPortalOmnibarSuggestionsConfiguration?
+    @State private var chromeStyle: BrowserChromeStyle
+    @State private var tabBarFontSize: CGFloat = GhosttyConfig.load(
+        globalFontMagnificationPercent: GlobalFontMagnification.storedPercent
+    ).surfaceTabBarFontSize
+
+    init(
+        panel: CEFBrowserPanel,
+        isFocused: Bool,
+        onRequestPanelFocus: @escaping () -> Void
+    ) {
+        self.panel = panel
+        self.isFocused = isFocused
+        self.onRequestPanelFocus = onRequestPanelFocus
+        self._chromeStyle = State(
+            initialValue: BrowserChromeStyle.resolve(
+                for: .light,
+                themeBackgroundColor: GhosttyBackgroundTheme.currentColor(),
+                drawsBackground: true
+            )
+        )
+    }
 
     var body: some View {
         VStack(spacing: 0) {
-            HStack(spacing: 6) {
-                Button {
-                    onRequestPanelFocus()
-                    panel.goBack()
-                } label: {
-                    Label(
-                        String(localized: "browser.goBack", defaultValue: "Go Back"),
-                        systemImage: "chevron.left"
-                    )
-                    .labelStyle(.iconOnly)
+            OmnibarPaneChrome(
+                panel: panel,
+                isFocused: isFocused,
+                chromeStyle: chromeStyle,
+                tabBarFontSize: tabBarFontSize,
+                accessorySpacing: 2,
+                onRequestPanelFocus: onRequestPanelFocus,
+                onReloadOrStop: {
+                    if panel.isLoading {
+                        panel.stopLoading()
+                    } else {
+                        panel.reload()
+                    }
+                },
+                onReload: panel.reload,
+                onHardReload: nil,
+                onAddressBarFocusStateChange: { _ in },
+                onChromeHeightChange: { _ in },
+                onSuggestionsPresentationChange: { configuration in
+                    suggestions = configuration
+                },
+                leadingAccessories: {
+                    EmptyView()
+                },
+                trailingAccessories: { _ in
+                    CEFExtensionActionBar()
                 }
-                .buttonStyle(.borderless)
-                .disabled(!panel.canGoBack)
-                .help(String(localized: "browser.goBack", defaultValue: "Go Back"))
+            )
 
-                Button {
-                    onRequestPanelFocus()
-                    panel.goForward()
-                } label: {
-                    Label(
-                        String(localized: "browser.goForward", defaultValue: "Go Forward"),
-                        systemImage: "chevron.right"
-                    )
-                    .labelStyle(.iconOnly)
-                }
-                .buttonStyle(.borderless)
-                .disabled(!panel.canGoForward)
-                .help(String(localized: "browser.goForward", defaultValue: "Go Forward"))
-
-                Button {
-                    onRequestPanelFocus()
-                    panel.reload()
-                } label: {
-                    Label(
-                        String(localized: "browser.reload", defaultValue: "Reload"),
-                        systemImage: "arrow.clockwise"
-                    )
-                    .labelStyle(.iconOnly)
-                }
-                .buttonStyle(.borderless)
-                .help(String(localized: "browser.reload", defaultValue: "Reload"))
-
-                TextField(
-                    String(localized: "cef.browser.address.placeholder", defaultValue: "Enter URL"),
-                    text: $panel.currentURL
-                )
-                .textFieldStyle(.roundedBorder)
-                .focused($isAddressFieldFocused)
-                .simultaneousGesture(TapGesture().onEnded {
-                    panel.setAddressFieldFocused(true)
-                    onRequestPanelFocus()
-                })
-                .onSubmit {
-                    panel.navigate(to: panel.currentURL)
-                }
-            }
-            .padding(.horizontal, 8)
-            .padding(.vertical, 6)
-
-            Divider()
-
-            CEFBrowserContainerRepresentable(
-                containerView: panel.containerView,
+            CEFBrowserHostRepresentable(
+                hostView: panel.hostView,
+                suggestions: suggestions,
                 onRequestPanelFocus: onRequestPanelFocus
             )
-                .frame(maxWidth: .infinity, maxHeight: .infinity)
-                .onAppear {
-                    panel.start(url: panel.currentURL)
-                }
-        }
-        .onChange(of: isAddressFieldFocused) { _, isFocused in
-            panel.setAddressFieldFocused(isFocused)
-        }
-        .onChange(of: isFocused) { _, isFocused in
-            if !isFocused {
-                isAddressFieldFocused = false
+            .frame(maxWidth: .infinity, maxHeight: .infinity)
+            .onAppear {
+                panel.start(url: panel.currentURL)
             }
+        }
+        .background(Color(nsColor: chromeStyle.backgroundColor))
+        .onAppear {
+            refreshChromeStyle()
+        }
+        .onChange(of: colorScheme) { _, _ in
+            refreshChromeStyle()
+        }
+        .onReceive(NotificationCenter.default.publisher(for: .ghosttyConfigDidReload)) { _ in
+            tabBarFontSize = GhosttyConfig.load(
+                globalFontMagnificationPercent: GlobalFontMagnification.storedPercent
+            ).surfaceTabBarFontSize
         }
     }
-}
 
-private struct CEFBrowserContainerRepresentable: NSViewRepresentable {
-    let containerView: CEFBrowserContainerView
-    let onRequestPanelFocus: () -> Void
-
-    func makeCoordinator() -> Coordinator {
-        Coordinator(containerView: containerView, onRequestPanelFocus: onRequestPanelFocus)
-    }
-
-    func makeNSView(context: Context) -> CEFBrowserContainerView {
-        containerView
-    }
-
-    func updateNSView(_ nsView: CEFBrowserContainerView, context: Context) {}
-
-    @MainActor
-    final class Coordinator {
-        private weak var containerView: CEFBrowserContainerView?
-        private let onRequestPanelFocus: () -> Void
-        private var eventMonitor: Any?
-
-        init(containerView: CEFBrowserContainerView, onRequestPanelFocus: @escaping () -> Void) {
-            self.containerView = containerView
-            self.onRequestPanelFocus = onRequestPanelFocus
-            eventMonitor = NSEvent.addLocalMonitorForEvents(matching: .leftMouseDown) { [weak self] event in
-                guard let self,
-                      let containerView = self.containerView,
-                      event.window === containerView.window,
-                      !containerView.isHiddenOrHasHiddenAncestor,
-                      self.eventTargetsContainer(event, containerView: containerView) else {
-                    return event
-                }
-
-                self.onRequestPanelFocus()
-                return event
-            }
-        }
-
-        private func eventTargetsContainer(
-            _ event: NSEvent,
-            containerView: CEFBrowserContainerView
-        ) -> Bool {
-            guard let contentView = event.window?.contentView else { return false }
-            let hitPoint = contentView.convert(event.locationInWindow, from: nil)
-            var hitView = contentView.hitTest(hitPoint)
-            while let view = hitView {
-                if view === containerView {
-                    return true
-                }
-                hitView = view.superview
-            }
-            return false
-        }
-
-        deinit {
-            if let eventMonitor {
-                NSEvent.removeMonitor(eventMonitor)
-            }
-        }
+    private func refreshChromeStyle() {
+        chromeStyle = BrowserChromeStyle.resolve(
+            for: colorScheme,
+            themeBackgroundColor: GhosttyBackgroundTheme.currentColor(),
+            drawsBackground: true
+        )
     }
 }
