+++ Sources/Panels/Omnibar/OmnibarInlineCompletionSupport.swift
@@ -1,0 +1,230 @@
+import AppKit
+import Foundation
+
+func staleOmnibarRemoteSuggestionsForDisplay(
+    query: String,
+    previousRemoteQuery: String,
+    previousRemoteSuggestions: [String],
+    allowsRemoteSuggestions: Bool = true,
+    limit: Int = 8
+) -> [String] {
+    guard allowsRemoteSuggestions else { return [] }
+    let trimmedQuery = query.trimmingCharacters(in: .whitespacesAndNewlines)
+    let trimmedPreviousQuery = previousRemoteQuery.trimmingCharacters(in: .whitespacesAndNewlines)
+    let loweredQuery = trimmedQuery.lowercased()
+    let loweredPreviousQuery = trimmedPreviousQuery.lowercased()
+    guard !trimmedQuery.isEmpty, !trimmedPreviousQuery.isEmpty else { return [] }
+    guard loweredQuery == loweredPreviousQuery || loweredQuery.hasPrefix(loweredPreviousQuery) || loweredPreviousQuery.hasPrefix(loweredQuery) else {
+        return []
+    }
+    guard !previousRemoteSuggestions.isEmpty else { return [] }
+    let sanitized = previousRemoteSuggestions.compactMap { raw -> String? in
+        let trimmed = raw.trimmingCharacters(in: .whitespacesAndNewlines)
+        guard !trimmed.isEmpty else { return nil }
+        return trimmed
+    }
+
+    if sanitized.isEmpty {
+        return []
+    }
+    return Array(sanitized.prefix(limit))
+}
+
+func omnibarInlineCompletionForDisplay(
+    typedText: String,
+    suggestions: [OmnibarSuggestion],
+    isFocused: Bool,
+    selectionRange: NSRange,
+    hasMarkedText: Bool
+) -> OmnibarInlineCompletion? {
+    guard isFocused else { return nil }
+    guard !hasMarkedText else { return nil }
+
+    let query = typedText.trimmingCharacters(in: .whitespacesAndNewlines)
+    guard !query.isEmpty else { return nil }
+    let loweredQuery = query.lowercased()
+    let typedIncludesScheme = loweredQuery.hasPrefix("https://") || loweredQuery.hasPrefix("http://")
+    let typedIncludesWWWPrefix = loweredQuery.hasPrefix("www.")
+    let queryCount = query.utf16.count
+
+    let urlCandidate = suggestions.first { suggestion in
+        guard let completion = omnibarSuggestionCompletion(for: suggestion) else { return false }
+        return omnibarSuggestionMatchesTypedPrefix(
+            typedText: query,
+            suggestionCompletion: completion,
+            suggestionTitle: omnibarSuggestionTitle(for: suggestion)
+        )
+    }
+    guard let candidate = urlCandidate else {
+        return nil
+    }
+
+    let acceptedText = candidate.completion
+    let displayText: String
+    if typedQueryHasExplicitPathOrQuery(query) {
+        if typedIncludesScheme {
+            displayText = acceptedText
+        } else if typedIncludesWWWPrefix {
+            displayText = stripHTTPSchemePrefix(acceptedText)
+        } else {
+            displayText = stripHTTPSchemeAndWWWPrefix(acceptedText)
+        }
+    } else if let hostOnlyDisplay = inlineCompletionHostDisplayText(
+        for: acceptedText,
+        typedIncludesScheme: typedIncludesScheme,
+        typedIncludesWWWPrefix: typedIncludesWWWPrefix
+    ) {
+        displayText = hostOnlyDisplay
+    } else {
+        if typedIncludesScheme {
+            displayText = acceptedText
+        } else if typedIncludesWWWPrefix {
+            displayText = stripHTTPSchemePrefix(acceptedText)
+        } else {
+            displayText = stripHTTPSchemeAndWWWPrefix(acceptedText)
+        }
+    }
+
+    guard omnibarSuggestionSupportsAutocompletion(query: query, suggestion: candidate) else { return nil }
+    // The display text must start with the typed query so the inline completion
+    // visually extends what the user typed rather than replacing it (e.g. a
+    // history entry matched via title "localhost:3000" whose URL is google.com
+    // should not replace a typed "l" with "g").
+    guard displayText.lowercased().hasPrefix(loweredQuery) else { return nil }
+    guard displayText.utf16.count > queryCount else {
+        return nil
+    }
+
+    let displayCount = displayText.utf16.count
+
+    let resolvedSelectionRange: NSRange = {
+        if selectionRange.location == NSNotFound {
+            return NSRange(location: queryCount, length: 0)
+        }
+        let clampedLocation = min(selectionRange.location, displayCount)
+        let remaining = max(0, displayCount - clampedLocation)
+        let clampedLength = min(selectionRange.length, remaining)
+        return NSRange(location: clampedLocation, length: clampedLength)
+    }()
+
+    let suffixRange = NSRange(location: queryCount, length: max(0, displayCount - queryCount))
+    let isCaretAtTypedBoundary = (resolvedSelectionRange.length == 0 && resolvedSelectionRange.location == queryCount)
+    let isSuffixSelection = NSEqualRanges(resolvedSelectionRange, suffixRange)
+    let isSelectAllSelection = (resolvedSelectionRange.location == 0 && resolvedSelectionRange.length == displayCount)
+    // Command+A can briefly report just the typed prefix selection before the full
+    // select-all range lands. Keep inline completion alive through that transition.
+    let typedPrefixSelection = NSRange(location: 0, length: queryCount)
+    let isTypedPrefixSelection = NSEqualRanges(resolvedSelectionRange, typedPrefixSelection)
+    guard isCaretAtTypedBoundary || isSuffixSelection || isSelectAllSelection || isTypedPrefixSelection else {
+        return nil
+    }
+
+    return OmnibarInlineCompletion(typedText: query, displayText: displayText, acceptedText: acceptedText)
+}
+
+func omnibarDesiredSelectionRangeForInlineCompletion(
+    currentSelection: NSRange,
+    inlineCompletion: OmnibarInlineCompletion
+) -> NSRange {
+    let typedCount = inlineCompletion.typedText.utf16.count
+    let typedPrefixSelection = NSRange(location: 0, length: typedCount)
+    let displayCount = inlineCompletion.displayText.utf16.count
+    let isSelectAll = currentSelection.location == 0 && currentSelection.length == displayCount
+    if isSelectAll ||
+        NSEqualRanges(currentSelection, inlineCompletion.suffixRange) ||
+        NSEqualRanges(currentSelection, typedPrefixSelection) {
+        return currentSelection
+    }
+    return inlineCompletion.suffixRange
+}
+
+func omnibarPublishedBufferTextForFieldChange(
+    fieldValue: String,
+    inlineCompletion: OmnibarInlineCompletion?,
+    selectionRange: NSRange?,
+    hasMarkedText: Bool
+) -> String {
+    guard !hasMarkedText else { return fieldValue }
+    guard let inlineCompletion else { return fieldValue }
+    guard fieldValue == inlineCompletion.displayText else { return fieldValue }
+    guard let selectionRange else { return inlineCompletion.typedText }
+
+    let typedCount = inlineCompletion.typedText.utf16.count
+    let displayCount = inlineCompletion.displayText.utf16.count
+    let typedPrefixSelection = NSRange(location: 0, length: typedCount)
+    let isCaretAtTypedBoundary = selectionRange.location == typedCount && selectionRange.length == 0
+    let isSuffixSelection = NSEqualRanges(selectionRange, inlineCompletion.suffixRange)
+    let isSelectAllSelection = selectionRange.location == 0 && selectionRange.length == displayCount
+    let isTypedPrefixSelection = NSEqualRanges(selectionRange, typedPrefixSelection)
+    if isCaretAtTypedBoundary || isSuffixSelection || isSelectAllSelection || isTypedPrefixSelection {
+        return inlineCompletion.typedText
+    }
+
+    return fieldValue
+}
+
+func omnibarInlineCompletionIfBufferMatchesTypedPrefix(
+    bufferText: String,
+    inlineCompletion: OmnibarInlineCompletion?
+) -> OmnibarInlineCompletion? {
+    guard let inlineCompletion else { return nil }
+    guard bufferText == inlineCompletion.typedText else { return nil }
+    return inlineCompletion
+}
+
+func omnibarPrefixAfterDeletingTrailingWord(from text: String) -> String {
+    let nsText = text as NSString
+    let fullRange = NSRange(location: 0, length: nsText.length)
+    var deletionStart = nsText.length
+    nsText.enumerateSubstrings(in: fullRange, options: [.byWords, .reverse]) { _, range, _, stop in
+        deletionStart = range.location
+        stop.pointee = true
+    }
+    return nsText.substring(to: deletionStart)
+}
+
+private func typedQueryHasExplicitPathOrQuery(_ typedQuery: String) -> Bool {
+    var normalized = typedQuery.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
+    if normalized.hasPrefix("https://") {
+        normalized.removeFirst("https://".count)
+    } else if normalized.hasPrefix("http://") {
+        normalized.removeFirst("http://".count)
+    }
+    return normalized.contains("/") || normalized.contains("?") || normalized.contains("#")
+}
+
+private func inlineCompletionHostDisplayText(
+    for acceptedText: String,
+    typedIncludesScheme: Bool,
+    typedIncludesWWWPrefix: Bool
+) -> String? {
+    guard let components = URLComponents(string: acceptedText),
+          var host = components.host?.trimmingCharacters(in: .whitespacesAndNewlines).lowercased(),
+          !host.isEmpty else {
+        return nil
+    }
+
+    if !typedIncludesWWWPrefix, host.hasPrefix("www.") {
+        host.removeFirst("www.".count)
+    }
+
+    let portSuffix: String
+    if let port = components.port {
+        let scheme = components.scheme?.lowercased()
+        let isDefaultPort =
+            (scheme == "https" && port == 443) ||
+            (scheme == "http" && port == 80)
+        portSuffix = isDefaultPort ? "" : ":\(port)"
+    } else {
+        portSuffix = ""
+    }
+
+    let hostWithPort = "\(host)\(portSuffix)"
+    if typedIncludesScheme {
+        let scheme = (components.scheme?.lowercased() == "http") ? "http" : "https"
+        return "\(scheme)://\(hostWithPort)"
+    }
+    return hostWithPort
+}
+
+
